Refactor shape tests files into a folder, add tests for missing methods (rotation, shearing, scale, polygon generation), move complex shapes into fixtures.py
