+++ tests/shape/fixtures.py
@@ -1,0 +1,14 @@
+FLATTEN_CIRCLE_ORIGINAL = "m 50 0 b 22 0 0 22 0 50 b 0 78 22 100 50 100 b 78 100 100 78 100 50 b 100 22 78 0 50 0"
+FLATTEN_CIRCLE_DEST = "m 50 0 l 48.692 0.016 47.393 0.064 46.104 0.143 44.824 0.254 43.555 0.395 42.296 0.567 41.049 0.769 39.812 1 38.588 1.261 37.375 1.55 36.175 1.868 34.988 2.215 33.814 2.589 32.654 2.991 31.507 3.419 30.375 3.875 29.257 4.357 28.155 4.865 27.068 5.398 25.996 5.957 24.941 6.541 23.902 7.149 22.88 7.781 21.875 8.438 20.888 9.117 19.918 9.82 18.967 10.545 18.035 11.293 17.122 12.063 16.228 12.854 15.354 13.667 14.5 14.5 13.667 15.354 12.854 16.228 12.063 17.122 11.293 18.035 10.545 18.967 9.82 19.918 9.117 20.888 8.438 21.875 7.781 22.88 7.149 23.902 6.541 24.941 5.957 25.996 5.398 27.068 4.865 28.155 4.357 29.257 3.875 30.375 3.419 31.507 2.991 32.654 2.589 33.814 2.215 34.988 1.868 36.175 1.55 37.375 1.261 38.588 1 39.812 0.769 41.049 0.567 42.296 0.395 43.555 0.254 44.824 0.143 46.104 0.064 47.393 0.016 48.692 0 50 0.016 51.308 0.064 52.607 0.143 53.896 0.254 55.176 0.395 56.445 0.567 57.704 0.769 58.951 1 60.188 1.261 61.412 1.55 62.625 1.868 63.825 2.215 65.012 2.589 66.186 2.991 67.346 3.419 68.493 3.875 69.625 4.357 70.743 4.865 71.845 5.398 72.932 5.957 74.004 6.541 75.059 7.149 76.098 7.781 77.12 8.438 78.125 9.117 79.112 9.82 80.082 10.545 81.033 11.293 81.965 12.063 82.878 12.854 83.772 13.667 84.646 14.5 85.5 15.354 86.333 16.228 87.146 17.122 87.937 18.035 88.707 18.967 89.455 19.918 90.18 20.888 90.883 21.875 91.562 22.88 92.219 23.902 92.851 24.941 93.459 25.996 94.043 27.068 94.602 28.155 95.135 29.257 95.643 30.375 96.125 31.507 96.581 32.654 97.009 33.814 97.411 34.988 97.785 36.175 98.132 37.375 98.45 38.588 98.739 39.812 99 41.049 99.231 42.296 99.433 43.555 99.605 44.824 99.746 46.104 99.857 47.393 99.936 48.692 99.984 50 100 51.308 99.984 52.607 99.936 53.896 99.857 55.176 99.746 56.445 99.605 57.704 99.433 58.951 99.231 60.188 99 61.412 98.739 62.625 98.45 63.825 98.132 65.012 97.785 66.186 97.411 67.346 97.009 68.493 96.581 69.625 96.125 70.743 95.643 71.845 95.135 72.932 94.602 74.004 94.043 75.059 93.459 76.098 92.851 77.12 92.219 78.125 91.562 79.112 90.883 80.082 90.18 81.033 89.455 81.965 88.707 82.878 87.937 83.772 87.146 84.646 86.333 85.5 85.5 86.333 84.646 87.146 83.772 87.937 82.878 88.707 81.965 89.455 81.033 90.18 80.082 90.883 79.112 91.562 78.125 92.219 77.12 92.851 76.098 93.459 75.059 94.043 74.004 94.602 72.932 95.135 71.845 95.643 70.743 96.125 69.625 96.581 68.493 97.009 67.346 97.411 66.186 97.785 65.012 98.132 63.825 98.45 62.625 98.739 61.412 99 60.188 99.231 58.951 99.433 57.704 99.605 56.445 99.746 55.176 99.857 53.896 99.936 52.607 99.984 51.308 100 50 99.984 48.692 99.936 47.393 99.857 46.104 99.746 44.824 99.605 43.555 99.433 42.296 99.231 41.049 99 39.812 98.739 38.588 98.45 37.375 98.132 36.175 97.785 34.988 97.411 33.814 97.009 32.654 96.581 31.507 96.125 30.375 95.643 29.257 95.135 28.155 94.602 27.068 94.043 25.996 93.459 24.941 92.851 23.902 92.219 22.88 91.562 21.875 90.883 20.888 90.18 19.918 89.455 18.967 88.707 18.035 87.937 17.122 87.146 16.228 86.333 15.354 85.5 14.5 84.646 13.667 83.772 12.854 82.878 12.063 81.965 11.293 81.033 10.545 80.082 9.82 79.112 9.117 78.125 8.438 77.12 7.781 76.098 7.149 75.059 6.541 74.004 5.957 72.932 5.398 71.845 4.865 70.743 4.357 69.625 3.875 68.493 3.419 67.346 2.991 66.186 2.589 65.012 2.215 63.825 1.868 62.625 1.55 61.412 1.261 60.188 1 58.951 0.769 57.704 0.567 56.445 0.395 55.176 0.254 53.896 0.143 52.607 0.064 51.308 0.016 50 0"
+
+FLATTEN_RECT_ORIGINAL = "m -100.5 0 l 100 0 b 100 100 -100 100 -100.5 0 c"
+FLATTEN_RECT_DEST = "m -100.5 0 l 100 0 99.964 2.325 99.855 4.614 99.676 6.866 99.426 9.082 99.108 11.261 98.723 13.403 98.271 15.509 97.754 17.578 97.173 19.611 96.528 21.606 95.822 23.566 95.056 25.488 94.23 27.374 93.345 29.224 92.403 31.036 91.405 32.812 90.352 34.552 89.246 36.255 88.086 37.921 86.876 39.551 85.614 41.144 84.304 42.7 82.945 44.22 81.54 45.703 80.088 47.15 78.592 48.56 77.053 49.933 75.471 51.27 73.848 52.57 72.184 53.833 70.482 55.06 68.742 56.25 66.965 57.404 65.153 58.521 63.307 59.601 61.427 60.645 59.515 61.652 57.572 62.622 55.599 63.556 53.598 64.453 51.569 65.314 49.514 66.138 47.433 66.925 45.329 67.676 43.201 68.39 41.052 69.067 38.882 69.708 36.692 70.312 34.484 70.88 32.259 71.411 27.762 72.363 23.209 73.169 18.61 73.828 13.975 74.341 9.311 74.707 4.629 74.927 -0.062 75 -4.755 74.927 -9.438 74.707 -14.103 74.341 -18.741 73.828 -23.343 73.169 -27.9 72.363 -32.402 71.411 -34.63 70.88 -36.841 70.312 -39.033 69.708 -41.207 69.067 -43.359 68.39 -45.49 67.676 -47.599 66.925 -49.683 66.138 -51.743 65.314 -53.776 64.453 -55.782 63.556 -57.759 62.622 -59.707 61.652 -61.624 60.645 -63.509 59.601 -65.361 58.521 -67.178 57.404 -68.961 56.25 -70.707 55.06 -72.415 53.833 -74.085 52.57 -75.714 51.27 -77.303 49.933 -78.85 48.56 -80.353 47.15 -81.811 45.703 -83.224 44.22 -84.59 42.7 -85.909 41.144 -87.178 39.551 -88.397 37.921 -89.564 36.255 -90.68 34.552 -91.741 32.812 -92.748 31.036 -93.699 29.224 -94.593 27.374 -95.428 25.488 -96.205 23.566 -96.92 21.606 -97.575 19.611 -98.166 17.578 -98.693 15.509 -99.156 13.403 -99.552 11.261 -99.881 9.082 -100.141 6.866 -100.332 4.614 -100.452 2.325 -100.5 0"
+
+FLATTEN_COMPLEX1_ORIGINAL = "m 35.281 142.422 b 27.812 142.422 22.375 140.406 18.984 136.375 15.594 132.328 13.891 125.734 13.891 116.562 l 13.891 80.453 6.875 80.453 6.875 63.672 13.891 63.672 13.891 41.594 32.188 41.594 32.188 63.672 49.859 63.672 49.859 80.453 32.188 80.453 32.188 107.828 b 32.188 110.531 32.203 112.891 32.25 114.906 32.297 116.922 32.578 118.734 33.078 120.344 33.578 121.953 34.453 123.219 35.688 124.156 36.938 125.109 38.734 125.578 41.125 125.578 42.094 125.578 43.375 125.297 44.984 124.75 46.578 124.203 47.688 123.688 48.281 123.234 l 49.859 123.234 49.859 140.219 b 47.891 140.906 45.781 141.453 43.531 141.828 41.281 142.219 38.531 142.422 35.281 142.422 l 35.281 142.422"
+FLATTEN_COMPLEX1_DEST = "m 35.281 142.422 l 34.587 142.416 33.904 142.398 33.234 142.369 32.575 142.328 31.929 142.274 31.294 142.209 30.671 142.133 30.06 142.044 29.462 141.944 28.875 141.831 28.299 141.707 27.736 141.572 27.185 141.424 26.646 141.264 26.119 141.093 25.603 140.91 25.1 140.715 24.608 140.509 24.129 140.29 23.661 140.06 23.206 139.818 22.762 139.564 22.33 139.298 21.911 139.02 21.503 138.731 21.107 138.43 20.723 138.117 20.352 137.792 19.992 137.456 19.644 137.107 19.308 136.747 18.984 136.375 18.671 135.988 18.368 135.586 18.075 135.17 17.792 134.738 17.519 134.291 17.255 133.83 17.002 133.353 16.758 132.862 16.524 132.355 16.301 131.834 16.087 131.297 15.883 130.746 15.689 130.18 15.505 129.598 15.33 129.002 15.166 128.39 15.012 127.764 14.868 127.123 14.733 126.466 14.609 125.795 14.494 125.108 14.389 124.406 14.21 122.958 14.071 121.45 13.971 119.881 13.911 118.252 13.891 116.562 13.891 80.453 6.875 80.453 6.875 63.672 13.891 63.672 13.891 41.594 32.188 41.594 32.188 63.672 49.859 63.672 49.859 80.453 32.188 80.453 32.188 107.828 32.25 114.906 32.279 115.652 32.329 116.38 32.401 117.088 32.494 117.777 32.549 118.115 32.609 118.447 32.674 118.775 32.744 119.099 32.82 119.417 32.901 119.731 32.987 120.04 33.078 120.344 33.176 120.642 33.229 120.787 33.283 120.931 33.34 121.073 33.399 121.213 33.46 121.351 33.523 121.487 33.589 121.621 33.656 121.752 33.726 121.882 33.798 122.01 33.872 122.136 33.948 122.26 34.027 122.382 34.107 122.502 34.19 122.62 34.275 122.736 34.362 122.85 34.451 122.962 34.543 123.073 34.636 123.181 34.732 123.287 34.829 123.391 34.929 123.494 35.031 123.594 35.136 123.693 35.242 123.789 35.35 123.884 35.461 123.976 35.573 124.067 35.688 124.156 35.807 124.244 35.929 124.329 36.054 124.411 36.182 124.491 36.314 124.567 36.449 124.641 36.587 124.712 36.729 124.78 36.873 124.846 37.021 124.908 37.173 124.968 37.327 125.025 37.485 125.079 37.646 125.13 37.811 125.179 37.979 125.224 38.15 125.267 38.324 125.307 38.502 125.345 38.684 125.379 38.868 125.411 39.056 125.44 39.443 125.49 39.842 125.528 40.256 125.556 40.684 125.573 41.125 125.578 41.217 125.577 41.31 125.575 41.503 125.565 41.703 125.548 41.91 125.526 42.125 125.496 42.348 125.46 42.577 125.418 42.814 125.369 43.059 125.314 43.311 125.252 43.839 125.11 44.396 124.943 44.984 124.75 45.559 124.546 46.088 124.346 46.571 124.15 47.008 123.957 47.209 123.862 47.398 123.769 47.575 123.677 47.74 123.585 47.893 123.496 48.034 123.407 48.101 123.363 48.164 123.32 48.224 123.277 48.281 123.234 49.859 123.234 49.859 140.219 49.114 140.47 48.356 140.707 47.585 140.931 46.801 141.141 46.003 141.335 45.192 141.515 44.368 141.679 43.531 141.828 42.664 141.966 41.75 142.086 40.789 142.188 39.781 142.272 38.726 142.337 37.625 142.384 35.281 142.422 35.281 142.422"
+
+FLATTEN_COMPLEX2_ORIGINAL = "m -550 269 b -544 269 -538 269 -532 268 b -522 266 -513 261 -507 253 b -503 249 -500 245 -497 240 b -496 239 -495 238 -492 238 b -487 239 -482 240 -476 241 b -473 242 -470 244 -467 247 b -462 252 -458 257 -453 263 b -447 270 -440 276 -433 282 b -419 294 -404 295 -388 288 b -386 287 -384 287 -382 285 b -376 282 -372 284 -367 289 b -362 295 -360 301 -357 309 b -354 319 -352 329 -354 339 b -354 342 -355 344 -356 346 b -365 364 -376 380 -394 390 b -408 398 -423 403 -439 406 b -461 409 -482 408 -504 406 b -505 405 -506 405 -506 403 b -508 389 -511 375 -517 361 b -523 345 -533 332 -549 323 b -556 320 -561 315 -562 307 b -562 306 -562 305 -562 304 b -562 300 -562 296 -562 292 m -492 39 b -491 39 -490 39 -490 39 b -481 43 -472 46 -465 52 b -463 54 -463 56 -465 58 b -481 83 -496 109 -505 137 b -508 147 -509 157 -507 168 b -507 170 -506 172 -505 174 b -498 188 -498 203 -504 218 b -512 240 -527 254 -551 257 b -557 258 -563 256 -568 254 b -582 248 -592 238 -601 227 b -624 198 -639 165 -648 129 b -648 127 -649 124 -649 122 b -649 115 -648 114 -642 111 b -635 108 -627 107 -620 106 b -583 102 -552 84 -526 60 b -519 54 -513 48 -506 44 b -501 41 -497 39 -492 39 m -181 85 b -181 88 -181 92 -181 95 b -181 97 -182 98 -184 98 b -192 98 -200 97 -208 95 b -209 94 -210 93 -210 92 b -221 73 -236 56 -254 44 b -270 33 -284 21 -300 11 b -304 9 -308 7 -312 5 b -319 3 -326 3 -333 5 b -353 10 -373 16 -392 24 b -403 29 -411 36 -416 47 b -416 48 -416 48 -417 49 b -419 55 -419 55 -426 52 b -441 45 -457 39 -473 32 b -474 32 -475 31 -476 31 b -478 30 -478 29 -477 27 b -454 -9 -424 -37 -379 -45 b -362 -49 -345 -54 -329 -62 b -318 -67 -309 -73 -301 -82 b -298 -84 -298 -84 -296 -81 b -286 -64 -274 -47 -261 -31 b -256 -25 -250 -20 -244 -15 b -222 4 -204 25 -190 49 b -183 60 -180 71 -181 85 m -325 301 b -342 300 -354 292 -363 277 b -365 274 -364 269 -363 265 b -360 257 -355 250 -351 242 b -346 233 -342 223 -343 213 b -343 208 -341 204 -338 201 b -330 193 -321 187 -312 181 b -294 170 -277 161 -258 152 b -245 147 -235 138 -227 127 b -222 122 -219 117 -215 111 b -213 108 -212 108 -211 111 b -204 125 -200 140 -204 156 b -208 173 -210 190 -212 207 b -213 221 -218 235 -226 247 b -227 250 -230 252 -232 254 b -257 272 -283 287 -312 298 b -316 300 -320 302 -325 301 m -415 82 b -416 90 -419 98 -421 105 b -423 110 -424 115 -426 120 b -430 130 -427 140 -423 149 b -416 162 -405 172 -393 181 b -387 185 -379 190 -372 194 b -359 200 -354 215 -360 228 b -365 237 -370 246 -372 255 b -375 270 -387 274 -398 277 b -407 279 -415 277 -422 271 b -432 263 -440 253 -448 244 b -454 237 -461 231 -470 229 b -476 228 -481 226 -486 224 b -488 223 -489 223 -488 220 b -487 214 -485 208 -484 202 b -483 194 -484 187 -487 180 b -495 164 -494 148 -488 132 b -481 113 -473 95 -463 78 b -460 73 -456 68 -452 64 b -448 61 -444 60 -439 61 b -431 62 -424 65 -418 70 b -417 71 -417 71 -416 73 m -158 304 b -159 319 -163 333 -169 347 b -170 350 -173 352 -176 353 b -184 358 -193 359 -202 358 b -205 358 -206 359 -206 361 b -208 365 -210 369 -212 372 b -218 379 -224 382 -233 380 b -245 377 -257 372 -268 366 b -277 361 -285 361 -294 366 b -303 371 -313 374 -324 375 b -333 376 -342 374 -349 367 b -350 365 -351 364 -349 362 b -339 349 -338 334 -342 319 b -342 318 -342 317 -342 316 b -343 310 -343 310 -337 311 b -332 312 -327 313 -322 313 b -308 315 -295 310 -283 305 b -263 297 -245 286 -229 271 b -218 261 -209 250 -205 234 b -202 243 -201 252 -197 259 b -195 265 -192 269 -187 272 b -179 276 -171 281 -162 286 b -159 288 -158 290 -158 294 b -158 297 -158 300 -158 304 m -197 109 b -175 110 -153 114 -131 122 b -116 127 -105 136 -98 150 b -91 165 -83 179 -74 193 b -73 194 -72 195 -70 199 b -69 201 -68 203 -71 204 b -81 209 -90 213 -98 218 b -102 221 -105 224 -107 228 b -114 242 -122 255 -134 264 b -136 265 -137 266 -139 267 b -148 273 -157 274 -167 268 b -187 254 -186 254 -190 231 b -193 219 -194 207 -196 194 b -197 184 -195 173 -193 163 b -188 144 -191 126 -197 109 m -349 192 b -347 174 -341 157 -333 141 b -327 127 -318 114 -311 101 b -305 92 -301 82 -299 72 b -297 63 -299 56 -305 50 b -311 43 -318 38 -325 34 b -328 32 -330 30 -332 28 b -334 27 -335 25 -334 23 b -334 21 -332 21 -330 21 b -322 20 -314 21 -306 25 b -280 40 -256 58 -234 79 b -223 89 -222 100 -231 112 b -239 122 -249 130 -260 137 b -272 144 -286 150 -298 157 b -315 166 -331 178 -346 191 b -347 191 -347 192 -349 192 m -163 442 b -176 447 -188 450 -201 452 b -218 456 -236 456 -253 452 b -281 445 -298 426 -307 400 b -309 397 -309 394 -311 390 b -311 389 -311 387 -309 387 b -301 384 -294 381 -286 378 b -282 377 -279 377 -276 379 b -266 383 -256 389 -246 392 b -242 393 -238 394 -233 394 b -227 395 -221 398 -216 401 b -197 412 -180 427 -163 442 m -378 37 b -380 48 -381 59 -380 71 b -377 93 -367 111 -346 122 b -344 123 -344 124 -344 126 b -351 143 -357 161 -363 178 b -364 180 -364 181 -366 180 b -380 173 -392 164 -402 153 b -407 147 -411 141 -412 133 b -413 130 -412 127 -412 125 b -409 113 -405 101 -401 89 b -398 80 -399 72 -403 64 b -406 58 -405 54 -402 50 b -397 41 -388 37 -379 35 b -377 35 -378 36 -378 37 m -653 401 b -643 393 -634 386 -624 378 b -611 368 -597 359 -581 352 b -570 347 -559 345 -546 348 b -530 351 -520 374 -534 388 b -547 400 -561 406 -578 408 b -599 410 -621 408 -642 403 b -645 402 -649 402 -653 401 m -405 474 b -400 463 -401 453 -405 442 b -408 432 -407 422 -404 412 b -403 409 -401 408 -398 407 b -384 401 -372 392 -362 381 b -360 378 -358 379 -356 380 b -346 388 -343 402 -347 416 b -350 426 -356 434 -365 439 b -379 447 -389 459 -400 470 b -401 472 -403 473 -404 475 b -404 475 -405 474 -405 474 m -315 68 b -315 76 -318 83 -321 89 b -325 97 -329 103 -334 109 b -336 111 -337 112 -339 111 b -345 107 -349 102 -352 98 b -354 96 -352 95 -351 94 b -343 87 -338 78 -334 69 b -332 63 -330 56 -330 50 b -330 47 -329 47 -327 48 b -319 52 -315 59 -315 68 m -360 83 b -365 74 -367 65 -366 55 b -366 48 -366 42 -365 35 b -364 32 -363 31 -361 32 b -359 33 -356 33 -354 34 b -346 38 -342 44 -343 53 b -345 65 -352 74 -360 83 m -562 292 b -561 296 -561 300 -562 304 b -563 300 -563 296 -562 292 m -158 304 b -160 300 -160 297 -158 294 b -157 297 -157 300 -158 304 m -415 82 b -417 79 -417 76 -416 73 b -415 76 -415 79 -415 82 m -378 37 b -378 36 -378 35 -379 35 b -378 35 -378 34 -377 35 b -377 36 -377 37 -378 37"
+FLATTEN_COMPLEX2_DEST = "m -550 269 l -545.5 268.984 -543.25 268.947 -541 268.875 -538.75 268.756 -536.5 268.578 -535.375 268.464 -534.25 268.33 -533.125 268.176 -532 268 -531.065 267.804 -530.137 267.59 -529.216 267.358 -528.301 267.109 -527.393 266.843 -526.494 266.559 -525.602 266.257 -524.719 265.938 -523.844 265.601 -522.979 265.246 -522.123 264.874 -521.277 264.484 -520.442 264.077 -519.617 263.652 -518.803 263.21 -518 262.75 -517.209 262.272 -516.43 261.777 -515.664 261.265 -514.91 260.734 -514.17 260.187 -513.443 259.621 -512.73 259.038 -512.031 258.438 -511.347 257.819 -510.678 257.184 -510.025 256.53 -509.387 255.859 -508.765 255.171 -508.16 254.465 -507.571 253.741 -507 253 -505.545 251.498 -504.172 249.984 -502.869 248.447 -501.625 246.875 -500.428 245.256 -499.266 243.578 -498.127 241.83 -497 240 -496.812 239.813 -496.621 239.627 -496.424 239.444 -496.323 239.354 -496.219 239.266 -496.112 239.178 -496.001 239.093 -495.888 239.009 -495.77 238.928 -495.647 238.848 -495.52 238.771 -495.388 238.697 -495.25 238.625 -495.106 238.556 -494.957 238.49 -494.8 238.428 -494.637 238.369 -494.466 238.314 -494.288 238.262 -494.101 238.215 -493.906 238.172 -493.703 238.133 -493.49 238.099 -493.267 238.069 -493.035 238.045 -492.793 238.026 -492.54 238.011 -492.275 238.003 -492 238 -484.375 239.5 -476 241 -475.719 241.097 -475.438 241.199 -475.156 241.308 -474.875 241.422 -474.594 241.542 -474.312 241.668 -473.75 241.938 -473.188 242.23 -472.625 242.547 -472.062 242.887 -471.5 243.25 -470.938 243.637 -470.375 244.047 -469.812 244.48 -469.25 244.938 -468.688 245.418 -468.125 245.922 -467.562 246.449 -467 247 -465.168 248.877 -463.406 250.766 -461.691 252.678 -460 254.625 -456.594 258.672 -453 263 -451.864 264.301 -450.705 265.58 -448.328 268.078 -445.881 270.506 -443.375 272.875 -440.822 275.197 -438.234 277.484 -433 282 -431.685 283.093 -430.363 284.122 -429.036 285.087 -427.703 285.99 -427.034 286.418 -426.364 286.831 -425.693 287.228 -425.02 287.61 -424.345 287.976 -423.669 288.327 -422.991 288.663 -422.312 288.984 -421.632 289.29 -420.95 289.581 -420.267 289.858 -419.582 290.119 -418.896 290.366 -418.208 290.597 -417.519 290.815 -416.828 291.018 -416.136 291.206 -415.442 291.38 -414.747 291.539 -414.051 291.685 -413.353 291.816 -412.653 291.933 -411.952 292.036 -411.25 292.125 -410.546 292.2 -409.841 292.261 -409.134 292.309 -408.426 292.343 -407.716 292.363 -407.005 292.369 -406.292 292.362 -405.578 292.342 -404.863 292.308 -404.146 292.261 -403.427 292.201 -402.707 292.127 -401.986 292.041 -401.263 291.941 -399.812 291.703 -398.356 291.414 -396.895 291.074 -395.427 290.684 -393.953 290.244 -392.474 289.755 -390.988 289.218 -389.497 288.633 -388 288 -387.812 287.909 -387.625 287.823 -387.438 287.743 -387.25 287.666 -386.875 287.523 -386.5 287.391 -385.75 287.139 -385.375 287.01 -385 286.875 -384.812 286.803 -384.625 286.728 -384.438 286.648 -384.25 286.564 -384.062 286.475 -383.875 286.381 -383.688 286.28 -383.5 286.172 -383.312 286.057 -383.125 285.934 -382.938 285.802 -382.75 285.662 -382.562 285.512 -382.375 285.352 -382.188 285.182 -382 285 -381.72 284.863 -381.443 284.733 -381.169 284.611 -380.898 284.496 -380.629 284.387 -380.363 284.286 -380.099 284.192 -379.838 284.105 -379.579 284.025 -379.323 283.952 -379.068 283.886 -378.816 283.827 -378.566 283.774 -378.318 283.728 -378.072 283.689 -377.828 283.656 -377.586 283.63 -377.345 283.611 -377.107 283.598 -376.869 283.591 -376.634 283.591 -376.4 283.597 -376.167 283.61 -375.936 283.629 -375.705 283.654 -375.477 283.685 -375.249 283.723 -375.022 283.766 -374.797 283.816 -374.572 283.871 -374.348 283.933 -374.125 284 -373.903 284.073 -373.681 284.152 -373.46 284.237 -373.24 284.328 -373.019 284.424 -372.8 284.526 -372.58 284.633 -372.361 284.746 -372.142 284.865 -371.924 284.988 -371.705 285.118 -371.486 285.252 -371.048 285.538 -370.609 285.844 -370.169 286.17 -369.727 286.517 -369.282 286.884 -368.834 287.27 -368.382 287.674 -367.927 288.098 -367 289 -366.54 289.563 -366.097 290.125 -365.67 290.689 -365.258 291.254 -364.861 291.82 -364.478 292.388 -364.108 292.958 -363.75 293.531 -363.069 294.686 -362.43 295.855 -361.825 297.042 -361.25 298.25 -360.698 299.481 -360.164 300.738 -359.125 303.344 -357 309 -355.928 312.75 -355.438 314.625 -354.984 316.5 -354.572 318.375 -354.205 320.25 -353.888 322.125 -353.625 324 -353.421 325.875 -353.342 326.812 -353.279 327.75 -353.234 328.688 -353.205 329.625 -353.195 330.562 -353.203 331.5 -353.23 332.438 -353.277 333.375 -353.344 334.312 -353.432 335.25 -353.54 336.188 -353.671 337.125 -353.824 338.062 -354 339 -354.003 339.278 -354.011 339.551 -354.026 339.818 -354.045 340.08 -354.069 340.337 -354.099 340.589 -354.133 340.836 -354.172 341.078 -354.215 341.316 -354.262 341.55 -354.314 341.78 -354.369 342.006 -354.428 342.228 -354.49 342.447 -354.556 342.663 -354.625 342.875 -354.771 343.291 -354.928 343.697 -355.093 344.094 -355.266 344.484 -355.444 344.868 -355.627 345.248 -356 346 -357.712 349.351 -359.479 352.648 -361.306 355.888 -363.203 359.062 -365.176 362.167 -367.232 365.195 -368.294 366.679 -369.38 368.142 -370.49 369.582 -371.625 371 -372.787 372.394 -373.976 373.765 -375.193 375.11 -376.439 376.43 -377.716 377.723 -379.023 378.989 -380.362 380.228 -381.734 381.438 -383.14 382.618 -384.58 383.769 -386.056 384.888 -387.568 385.977 -389.118 387.033 -390.706 388.056 -392.333 389.045 -394 390 -396.637 391.465 -399.297 392.861 -401.98 394.19 -404.688 395.453 -407.418 396.652 -410.172 397.787 -412.949 398.861 -415.75 399.875 -418.574 400.83 -421.422 401.729 -424.293 402.571 -427.188 403.359 -430.105 404.095 -433.047 404.779 -436.012 405.414 -439 406 -443.114 406.516 -447.207 406.943 -451.283 407.285 -455.344 407.547 -459.393 407.732 -463.434 407.846 -471.5 407.875 -479.566 407.67 -487.656 407.266 -495.793 406.697 -504 406 -504.094 405.909 -504.187 405.823 -504.28 405.743 -504.373 405.666 -504.465 405.593 -504.556 405.523 -504.734 405.391 -504.907 405.264 -504.991 405.201 -505.072 405.139 -505.152 405.075 -505.229 405.01 -505.266 404.977 -505.303 404.944 -505.339 404.91 -505.375 404.875 -505.41 404.839 -505.444 404.803 -505.477 404.766 -505.51 404.728 -505.541 404.689 -505.572 404.648 -505.602 404.607 -505.631 404.564 -505.659 404.521 -505.686 404.475 -505.712 404.429 -505.738 404.381 -505.762 404.331 -505.785 404.28 -505.807 404.227 -505.828 404.172 -505.848 404.115 -505.867 404.057 -505.885 403.996 -505.901 403.934 -505.916 403.869 -505.931 403.802 -505.943 403.733 -505.955 403.662 -505.965 403.588 -505.974 403.512 -505.982 403.434 -505.989 403.352 -505.994 403.268 -505.997 403.182 -506 403 -506.801 397.75 -507.719 392.5 -508.777 387.25 -510 382 -510.68 379.375 -511.41 376.75 -512.193 374.125 -513.031 371.5 -513.928 368.875 -514.887 366.25 -515.91 363.625 -517 361 -517.574 359.509 -518.172 358.035 -518.795 356.58 -519.441 355.143 -520.113 353.724 -520.81 352.323 -521.533 350.941 -522.281 349.578 -523.056 348.234 -523.858 346.909 -524.687 345.604 -525.543 344.318 -526.427 343.052 -527.339 341.806 -528.28 340.581 -529.25 339.375 -530.249 338.19 -531.278 337.026 -532.337 335.882 -533.426 334.76 -534.546 333.659 -535.697 332.579 -536.879 331.521 -538.094 330.484 -539.34 329.47 -540.62 328.478 -541.932 327.508 -543.277 326.561 -544.657 325.636 -546.07 324.734 -547.518 323.855 -549 323 -549.65 322.713 -550.289 322.414 -550.914 322.103 -551.527 321.779 -552.127 321.443 -552.713 321.095 -553.286 320.734 -553.844 320.359 -554.387 319.972 -554.916 319.571 -555.429 319.157 -555.926 318.729 -556.407 318.286 -556.872 317.83 -557.319 317.36 -557.537 317.119 -557.75 316.875 -557.959 316.627 -558.163 316.375 -558.363 316.12 -558.558 315.861 -558.749 315.598 -558.935 315.332 -559.116 315.061 -559.293 314.787 -559.465 314.509 -559.632 314.227 -559.794 313.941 -559.952 313.652 -560.104 313.358 -560.252 313.06 -560.394 312.759 -560.531 312.453 -560.663 312.143 -560.79 311.83 -560.912 311.512 -561.029 311.19 -561.14 310.864 -561.246 310.534 -561.346 310.2 -561.441 309.861 -561.531 309.519 -561.615 309.172 -561.693 308.821 -561.766 308.465 -561.895 307.741 -562 307 -562 304 -562 292 m -492 39 l -490 39 -486.629 40.461 -483.281 41.875 -479.98 43.289 -478.355 44.011 -476.75 44.75 -475.168 45.513 -473.613 46.305 -472.087 47.132 -470.594 48 -469.86 48.451 -469.135 48.915 -468.42 49.392 -467.715 49.883 -467.02 50.388 -466.335 50.909 -465.662 51.446 -465 52 -464.818 52.188 -464.648 52.375 -464.568 52.469 -464.49 52.562 -464.416 52.656 -464.344 52.75 -464.275 52.844 -464.209 52.938 -464.146 53.031 -464.086 53.125 -464.029 53.219 -463.975 53.312 -463.923 53.406 -463.875 53.5 -463.83 53.594 -463.787 53.688 -463.748 53.781 -463.711 53.875 -463.677 53.969 -463.646 54.062 -463.619 54.156 -463.594 54.25 -463.572 54.344 -463.553 54.438 -463.537 54.531 -463.523 54.625 -463.513 54.719 -463.506 54.812 -463.501 54.906 -463.5 55 -463.501 55.094 -463.506 55.188 -463.513 55.281 -463.523 55.375 -463.537 55.469 -463.553 55.562 -463.572 55.656 -463.594 55.75 -463.619 55.844 -463.646 55.938 -463.677 56.031 -463.711 56.125 -463.748 56.219 -463.787 56.312 -463.83 56.406 -463.875 56.5 -463.923 56.594 -463.975 56.688 -464.029 56.781 -464.086 56.875 -464.146 56.969 -464.209 57.062 -464.275 57.156 -464.344 57.25 -464.416 57.344 -464.49 57.438 -464.568 57.531 -464.648 57.625 -464.818 57.812 -465 58 -470.943 67.424 -476.734 76.953 -482.314 86.6 -487.625 96.375 -490.161 101.315 -492.607 106.291 -494.957 111.305 -497.203 116.359 -499.338 121.454 -501.354 126.592 -503.243 131.773 -505 137 -505.539 138.875 -506.029 140.752 -506.47 142.632 -506.859 144.516 -507.196 146.406 -507.479 148.303 -507.705 150.209 -507.875 152.125 -507.986 154.053 -508.037 155.994 -508.027 157.95 -507.953 159.922 -507.815 161.911 -507.611 163.92 -507.34 165.949 -507 168 -506.997 168.188 -506.989 168.375 -506.974 168.562 -506.955 168.75 -506.931 168.938 -506.901 169.125 -506.867 169.312 -506.828 169.5 -506.785 169.688 -506.738 169.875 -506.686 170.062 -506.631 170.25 -506.572 170.438 -506.51 170.625 -506.375 171 -506.229 171.375 -506.072 171.75 -505.907 172.125 -505.734 172.5 -505.373 173.25 -505 174 -504.364 175.315 -503.769 176.636 -503.215 177.963 -502.701 179.295 -502.228 180.632 -501.794 181.974 -501.401 183.321 -501.047 184.672 -500.733 186.028 -500.458 187.387 -500.222 188.751 -500.025 190.119 -499.868 191.491 -499.748 192.865 -499.668 194.244 -499.625 195.625 -499.621 197.009 -499.654 198.396 -499.725 199.786 -499.834 201.178 -499.98 202.572 -500.163 203.968 -500.384 205.366 -500.641 206.766 -500.934 208.167 -501.264 209.569 -501.631 210.973 -502.033 212.377 -502.472 213.782 -502.946 215.188 -503.455 216.594 -504 218 -504.771 220.039 -505.583 222.031 -506.436 223.974 -507.332 225.869 -508.27 227.715 -509.251 229.511 -510.276 231.258 -511.344 232.953 -512.456 234.597 -513.612 236.19 -514.207 236.966 -514.813 237.73 -515.43 238.48 -516.059 239.217 -516.699 239.94 -517.35 240.65 -518.013 241.347 -518.687 242.03 -519.373 242.699 -520.07 243.355 -520.779 243.997 -521.5 244.625 -522.232 245.239 -522.977 245.839 -523.733 246.425 -524.5 246.997 -525.28 247.555 -526.072 248.099 -526.876 248.628 -527.691 249.143 -528.519 249.643 -529.359 250.129 -530.211 250.6 -531.076 251.056 -531.952 251.498 -532.841 251.925 -533.743 252.337 -534.656 252.734 -535.582 253.117 -536.521 253.484 -537.472 253.835 -538.436 254.172 -539.412 254.494 -540.402 254.8 -541.403 255.09 -542.418 255.365 -544.486 255.869 -546.605 256.309 -548.776 256.687 -551 257 -551.562 257.085 -552.125 257.153 -552.687 257.205 -553.248 257.24 -553.809 257.26 -554.368 257.266 -554.927 257.257 -555.484 257.234 -556.04 257.199 -556.594 257.15 -557.147 257.09 -557.697 257.018 -558.245 256.935 -558.791 256.841 -559.335 256.738 -559.875 256.625 -560.947 256.374 -562.006 256.092 -563.05 255.783 -564.078 255.453 -565.089 255.105 -566.08 254.744 -568 254 -569.301 253.426 -570.579 252.829 -571.835 252.21 -573.068 251.568 -574.281 250.906 -575.473 250.223 -576.645 249.52 -577.797 248.797 -578.93 248.055 -580.045 247.295 -581.141 246.516 -582.221 245.721 -583.283 244.908 -584.329 244.079 -586.375 242.375 -588.362 240.612 -590.295 238.795 -592.178 236.928 -594.016 235.016 -595.812 233.062 -597.572 231.072 -601 227 -605.219 221.516 -609.254 215.939 -613.107 210.272 -616.781 204.516 -620.28 198.671 -623.605 192.74 -626.761 186.724 -629.75 180.625 -632.575 174.444 -635.238 168.182 -637.744 161.841 -640.094 155.422 -642.292 148.927 -644.34 142.357 -646.242 135.715 -648 129 -648.003 128.81 -648.011 128.614 -648.025 128.413 -648.043 128.207 -648.066 127.997 -648.092 127.783 -648.156 127.344 -648.232 126.893 -648.316 126.434 -648.5 125.5 -648.684 124.566 -648.768 124.107 -648.844 123.656 -648.908 123.217 -648.934 123.003 -648.957 122.793 -648.975 122.587 -648.989 122.386 -648.997 122.19 -649 122 -648.987 120.756 -648.97 120.183 -648.945 119.641 -648.911 119.128 -648.868 118.643 -648.815 118.184 -648.784 117.964 -648.75 117.75 -648.713 117.542 -648.674 117.34 -648.631 117.143 -648.585 116.951 -648.536 116.765 -648.483 116.583 -648.427 116.407 -648.367 116.234 -648.304 116.067 -648.237 115.903 -648.166 115.744 -648.091 115.588 -648.012 115.436 -647.929 115.287 -647.842 115.142 -647.75 115 -647.654 114.861 -647.554 114.724 -647.449 114.59 -647.339 114.459 -647.224 114.33 -647.105 114.202 -646.981 114.077 -646.852 113.953 -646.717 113.831 -646.578 113.71 -646.433 113.59 -646.282 113.471 -646.126 113.352 -645.965 113.235 -645.625 113 -645.262 112.765 -644.874 112.529 -644.462 112.29 -644.023 112.047 -643.067 111.541 -642 111 -641.341 110.725 -640.676 110.46 -639.332 109.965 -637.97 109.51 -636.594 109.094 -635.206 108.712 -633.809 108.363 -632.406 108.043 -631 107.75 -628.191 107.23 -625.406 106.781 -620 106 -616.549 105.584 -613.133 105.088 -609.751 104.512 -606.404 103.859 -603.092 103.13 -599.814 102.326 -596.57 101.449 -593.359 100.5 -590.183 99.481 -587.04 98.393 -583.93 97.237 -580.854 96.016 -577.81 94.73 -574.799 93.381 -571.821 91.97 -568.875 90.5 -565.961 88.971 -563.08 87.385 -560.23 85.743 -557.412 84.047 -554.626 82.298 -551.87 80.498 -549.146 78.648 -546.453 76.75 -541.159 72.814 -535.986 68.703 -530.934 64.428 -526 60 -523.418 57.754 -520.906 55.531 -518.441 53.355 -516 51.25 -513.559 49.238 -512.331 48.275 -511.094 47.344 -509.845 46.448 -508.582 45.59 -507.301 44.773 -506 44 -504.168 42.924 -503.28 42.425 -502.406 41.953 -501.544 41.511 -500.691 41.1 -499.844 40.72 -499.422 40.543 -499 40.375 -498.578 40.215 -498.156 40.065 -497.733 39.923 -497.309 39.791 -496.883 39.668 -496.456 39.555 -496.026 39.452 -495.594 39.359 -495.159 39.277 -494.72 39.204 -494.278 39.143 -493.832 39.092 -493.382 39.052 -492.926 39.023 -492.466 39.006 -492 39 m -181 85 l -181 95 -181.003 95.185 -181.012 95.363 -181.018 95.45 -181.026 95.536 -181.036 95.62 -181.047 95.703 -181.059 95.784 -181.073 95.864 -181.089 95.943 -181.105 96.02 -181.124 96.095 -181.144 96.169 -181.165 96.241 -181.188 96.312 -181.212 96.382 -181.237 96.45 -181.264 96.517 -181.293 96.582 -181.323 96.646 -181.354 96.708 -181.387 96.769 -181.422 96.828 -181.458 96.886 -181.495 96.942 -181.534 96.997 -181.574 97.051 -181.616 97.103 -181.659 97.153 -181.704 97.202 -181.75 97.25 -181.798 97.296 -181.847 97.341 -181.897 97.384 -181.949 97.426 -182.003 97.466 -182.058 97.505 -182.114 97.542 -182.172 97.578 -182.231 97.613 -182.292 97.646 -182.354 97.677 -182.418 97.707 -182.483 97.736 -182.55 97.763 -182.618 97.788 -182.688 97.812 -182.759 97.835 -182.831 97.856 -182.905 97.876 -182.98 97.895 -183.057 97.911 -183.136 97.927 -183.216 97.941 -183.297 97.953 -183.38 97.964 -183.464 97.974 -183.55 97.982 -183.637 97.988 -183.815 97.997 -184 98 -187 97.953 -190 97.812 -193 97.578 -196 97.25 -199 96.828 -202 96.312 -205 95.703 -208 95 -208.373 94.625 -208.734 94.25 -208.907 94.062 -209.072 93.875 -209.229 93.688 -209.303 93.594 -209.375 93.5 -209.444 93.406 -209.51 93.312 -209.572 93.219 -209.631 93.125 -209.686 93.031 -209.738 92.938 -209.762 92.891 -209.785 92.844 -209.807 92.797 -209.828 92.75 -209.848 92.703 -209.867 92.656 -209.885 92.609 -209.901 92.562 -209.916 92.516 -209.931 92.469 -209.943 92.422 -209.955 92.375 -209.965 92.328 -209.974 92.281 -209.982 92.234 -209.989 92.188 -209.994 92.141 -209.997 92.094 -209.999 92.047 -210 92 -212.109 88.462 -214.311 84.975 -216.603 81.543 -218.984 78.172 -221.454 74.865 -224.01 71.627 -226.651 68.462 -229.375 65.375 -232.181 62.37 -235.068 59.451 -238.034 56.623 -241.078 53.891 -244.198 51.258 -247.393 48.729 -250.66 46.308 -254 44 -259.914 39.834 -265.688 35.609 -277 27.125 -282.633 22.936 -288.312 18.828 -294.086 14.838 -297.022 12.898 -300 11 -312 5 -313.312 4.648 -314.625 4.344 -315.938 4.086 -316.594 3.975 -317.25 3.875 -317.906 3.787 -318.562 3.711 -319.219 3.646 -319.875 3.594 -320.531 3.553 -321.188 3.523 -321.844 3.506 -322.5 3.5 -323.156 3.506 -323.812 3.523 -324.469 3.553 -325.125 3.594 -325.781 3.646 -326.438 3.711 -327.094 3.787 -327.75 3.875 -328.406 3.975 -329.062 4.086 -330.375 4.344 -331.688 4.648 -333 5 -347.984 8.953 -355.447 11.1 -362.875 13.375 -370.256 15.791 -377.578 18.359 -384.83 21.092 -392 24 -394.027 24.961 -395.015 25.461 -395.984 25.973 -396.937 26.498 -397.871 27.037 -398.788 27.589 -399.688 28.156 -400.569 28.738 -401.434 29.334 -402.28 29.946 -403.109 30.574 -403.921 31.218 -404.715 31.878 -405.491 32.556 -406.25 33.25 -406.991 33.962 -407.715 34.692 -408.421 35.44 -409.109 36.207 -409.78 36.993 -410.434 37.798 -411.069 38.623 -411.688 39.469 -412.288 40.335 -412.871 41.221 -413.437 42.129 -413.984 43.059 -414.515 44.01 -415.027 44.984 -415.522 45.98 -416 47 -416 47.176 -416.002 47.332 -416.004 47.403 -416.007 47.47 -416.01 47.534 -416.016 47.594 -416.022 47.651 -416.026 47.679 -416.031 47.706 -416.035 47.732 -416.041 47.758 -416.046 47.784 -416.053 47.809 -416.06 47.833 -416.067 47.858 -416.075 47.882 -416.084 47.906 -416.093 47.929 -416.103 47.953 -416.114 47.977 -416.125 48 -416.137 48.023 -416.15 48.047 -416.164 48.071 -416.178 48.094 -416.193 48.118 -416.209 48.142 -416.226 48.167 -416.244 48.191 -416.283 48.242 -416.325 48.294 -416.371 48.349 -416.422 48.406 -416.536 48.53 -416.67 48.668 -417 49 -417.354 50.055 -417.674 50.975 -417.825 51.384 -417.973 51.762 -418.12 52.108 -418.193 52.269 -418.266 52.422 -418.339 52.567 -418.413 52.705 -418.488 52.836 -418.564 52.959 -418.602 53.017 -418.641 53.074 -418.68 53.129 -418.719 53.182 -418.759 53.234 -418.799 53.283 -418.84 53.331 -418.881 53.377 -418.922 53.421 -418.965 53.463 -419.007 53.504 -419.051 53.543 -419.095 53.58 -419.139 53.615 -419.184 53.649 -419.23 53.681 -419.277 53.711 -419.324 53.74 -419.372 53.766 -419.421 53.791 -419.471 53.815 -419.521 53.837 -419.573 53.857 -419.625 53.875 -419.678 53.892 -419.732 53.907 -419.787 53.92 -419.844 53.932 -419.901 53.942 -419.959 53.951 -420.018 53.958 -420.078 53.964 -420.14 53.968 -420.202 53.97 -420.266 53.971 -420.331 53.97 -420.397 53.967 -420.465 53.964 -420.533 53.958 -420.604 53.951 -420.675 53.943 -420.748 53.933 -420.897 53.908 -421.052 53.878 -421.214 53.842 -421.381 53.8 -421.555 53.753 -421.735 53.699 -421.922 53.641 -422.317 53.507 -422.741 53.351 -423.686 52.979 -426 52 -431.67 49.418 -437.422 46.906 -449.125 42 -461.016 37.094 -473 32 -473.047 31.999 -473.094 31.997 -473.141 31.994 -473.188 31.989 -473.234 31.983 -473.281 31.975 -473.328 31.967 -473.375 31.957 -473.422 31.946 -473.469 31.934 -473.516 31.922 -473.562 31.908 -473.609 31.893 -473.656 31.877 -473.75 31.844 -473.844 31.807 -473.938 31.768 -474.031 31.727 -474.125 31.684 -474.312 31.593 -474.5 31.5 -474.688 31.407 -474.875 31.316 -474.969 31.273 -475.062 31.232 -475.156 31.193 -475.25 31.156 -475.344 31.123 -475.391 31.107 -475.438 31.092 -475.484 31.078 -475.531 31.066 -475.578 31.054 -475.625 31.043 -475.672 31.033 -475.719 31.025 -475.766 31.017 -475.812 31.011 -475.859 31.006 -475.906 31.003 -475.953 31.001 -476 31 -476.182 30.906 -476.352 30.812 -476.511 30.718 -476.586 30.671 -476.658 30.623 -476.728 30.575 -476.795 30.527 -476.859 30.479 -476.921 30.431 -476.98 30.382 -477.036 30.333 -477.09 30.284 -477.141 30.234 -477.189 30.184 -477.235 30.134 -477.279 30.083 -477.32 30.032 -477.339 30.006 -477.358 29.98 -477.376 29.954 -477.394 29.928 -477.411 29.902 -477.428 29.875 -477.444 29.849 -477.459 29.822 -477.474 29.795 -477.488 29.769 -477.501 29.741 -477.514 29.714 -477.527 29.687 -477.538 29.659 -477.55 29.632 -477.56 29.604 -477.57 29.576 -477.58 29.548 -477.589 29.519 -477.597 29.491 -477.605 29.462 -477.612 29.433 -477.619 29.404 -477.625 29.375 -477.631 29.346 -477.636 29.316 -477.64 29.286 -477.644 29.256 -477.647 29.226 -477.65 29.196 -477.653 29.165 -477.655 29.135 -477.657 29.072 -477.657 29.009 -477.655 28.946 -477.65 28.881 -477.644 28.815 -477.636 28.749 -477.626 28.681 -477.614 28.613 -477.6 28.543 -477.584 28.472 -477.566 28.401 -477.547 28.328 -477.525 28.254 -477.502 28.179 -477.45 28.026 -477.392 27.868 -477.326 27.705 -477.254 27.537 -477.176 27.364 -477 27 -472.604 20.347 -470.34 17.096 -468.031 13.898 -465.676 10.757 -463.271 7.673 -460.818 4.649 -458.312 1.688 -455.755 -1.21 -453.143 -4.04 -450.475 -6.802 -447.75 -9.492 -444.967 -12.109 -442.123 -14.651 -439.218 -17.116 -436.25 -19.5 -433.218 -21.802 -430.119 -24.021 -426.953 -26.152 -423.719 -28.195 -420.414 -30.148 -417.037 -32.007 -413.587 -33.771 -410.062 -35.438 -406.462 -37.005 -402.783 -38.47 -399.026 -39.831 -395.188 -41.086 -391.267 -42.233 -387.264 -43.269 -383.175 -44.192 -379 -45 -372.627 -46.551 -366.266 -48.219 -359.928 -50.027 -353.625 -52 -347.369 -54.16 -344.262 -55.318 -341.172 -56.531 -338.099 -57.803 -335.045 -59.137 -332.011 -60.535 -329 -62 -326.961 -62.95 -324.967 -63.926 -323.017 -64.931 -321.109 -65.969 -319.243 -67.042 -317.416 -68.152 -315.627 -69.304 -313.875 -70.5 -312.158 -71.743 -310.475 -73.035 -308.823 -74.38 -307.203 -75.781 -305.612 -77.241 -304.049 -78.762 -302.512 -80.347 -301 -82 -300.471 -82.351 -300.006 -82.654 -299.794 -82.787 -299.596 -82.907 -299.41 -83.015 -299.321 -83.064 -299.234 -83.109 -299.151 -83.152 -299.069 -83.191 -298.991 -83.226 -298.914 -83.259 -298.839 -83.287 -298.767 -83.313 -298.731 -83.324 -298.696 -83.335 -298.661 -83.345 -298.627 -83.354 -298.593 -83.362 -298.56 -83.369 -298.527 -83.375 -298.494 -83.38 -298.462 -83.385 -298.43 -83.388 -298.398 -83.391 -298.366 -83.393 -298.335 -83.394 -298.305 -83.394 -298.274 -83.393 -298.244 -83.391 -298.214 -83.388 -298.184 -83.385 -298.154 -83.38 -298.125 -83.375 -298.096 -83.369 -298.067 -83.361 -298.038 -83.353 -298.009 -83.344 -297.981 -83.334 -297.952 -83.323 -297.924 -83.311 -297.895 -83.299 -297.867 -83.285 -297.839 -83.27 -297.811 -83.255 -297.783 -83.238 -297.754 -83.22 -297.726 -83.202 -297.698 -83.183 -297.67 -83.162 -297.642 -83.141 -297.613 -83.118 -297.585 -83.095 -297.557 -83.071 -297.499 -83.019 -297.442 -82.964 -297.383 -82.905 -297.324 -82.842 -297.264 -82.774 -297.203 -82.703 -297.141 -82.628 -297.078 -82.549 -296.947 -82.378 -296.81 -82.19 -296.666 -81.986 -296.353 -81.528 -296 -81 -292.158 -74.627 -288.141 -68.266 -283.959 -61.928 -279.625 -55.625 -275.15 -49.369 -270.547 -43.172 -265.826 -37.045 -261 -31 -260.051 -29.886 -259.08 -28.795 -258.089 -27.724 -257.078 -26.672 -255.006 -24.619 -252.875 -22.625 -250.697 -20.678 -248.484 -18.766 -244 -15 -239.922 -11.414 -235.938 -7.779 -232.047 -4.095 -228.25 -0.359 -224.547 3.429 -220.938 7.271 -217.422 11.17 -214 15.125 -210.672 19.139 -207.438 23.213 -204.297 27.348 -201.25 31.547 -198.297 35.81 -195.438 40.139 -192.672 44.535 -190 49 -188.734 51.063 -188.137 52.096 -187.562 53.131 -187.012 54.168 -186.484 55.207 -185.98 56.25 -185.5 57.297 -185.043 58.348 -184.609 59.404 -184.199 60.466 -183.812 61.533 -183.449 62.607 -183.109 63.689 -182.793 64.778 -182.5 65.875 -182.23 66.981 -181.984 68.096 -181.762 69.222 -181.562 70.357 -181.387 71.504 -181.234 72.662 -181.105 73.833 -181 75.016 -180.918 76.212 -180.859 77.422 -180.824 78.646 -180.812 79.885 -180.824 81.139 -180.859 82.409 -180.918 83.696 -181 85 m -325 301 l -326.579 300.886 -328.129 300.73 -329.651 300.534 -331.145 300.297 -332.61 300.019 -334.048 299.699 -335.459 299.339 -336.844 298.938 -338.202 298.495 -339.534 298.012 -340.84 297.487 -342.121 296.922 -343.377 296.315 -344.609 295.668 -345.816 294.979 -347 294.25 -348.16 293.479 -349.297 292.668 -350.412 291.815 -351.504 290.922 -352.574 289.987 -353.623 289.012 -354.65 287.995 -355.656 286.938 -356.642 285.839 -357.608 284.699 -358.554 283.519 -359.48 282.297 -360.388 281.034 -361.277 279.73 -362.147 278.386 -363 277 -363.092 276.858 -363.179 276.713 -363.262 276.565 -363.341 276.415 -363.415 276.262 -363.486 276.106 -363.553 275.948 -363.615 275.787 -363.674 275.624 -363.729 275.459 -363.781 275.291 -363.828 275.121 -363.873 274.949 -363.913 274.776 -363.95 274.6 -363.984 274.422 -364.015 274.242 -364.042 274.061 -364.067 273.878 -364.088 273.693 -364.121 273.319 -364.143 272.939 -364.153 272.555 -364.154 272.165 -364.144 271.772 -364.125 271.375 -364.097 270.975 -364.061 270.573 -364.018 270.169 -363.967 269.764 -363.846 268.951 -363.703 268.141 -363.543 267.336 -363.369 266.541 -363 265 -362.415 263.511 -361.787 262.043 -361.121 260.592 -360.422 259.156 -359.693 257.732 -358.939 256.316 -357.375 253.5 -354.141 247.844 -352.541 244.957 -351.761 243.489 -351 242 -350.075 240.301 -349.18 238.58 -348.319 236.839 -347.5 235.078 -346.728 233.3 -346.008 231.506 -345.347 229.697 -345.04 228.788 -344.75 227.875 -344.478 226.96 -344.224 226.041 -343.989 225.12 -343.773 224.197 -343.579 223.272 -343.405 222.344 -343.254 221.415 -343.125 220.484 -343.02 219.552 -342.938 218.618 -342.882 217.684 -342.852 216.748 -342.847 215.812 -342.87 214.875 -342.921 213.937 -343 213 -342.994 212.534 -342.977 212.074 -342.948 211.62 -342.908 211.172 -342.857 210.729 -342.796 210.293 -342.723 209.862 -342.641 209.438 -342.548 209.019 -342.445 208.605 -342.332 208.198 -342.209 207.797 -342.077 207.401 -341.935 207.012 -341.785 206.628 -341.625 206.25 -341.457 205.878 -341.28 205.512 -341.094 205.151 -340.9 204.797 -340.699 204.448 -340.489 204.105 -340.272 203.769 -340.047 203.438 -339.815 203.112 -339.575 202.793 -339.329 202.479 -339.076 202.172 -338.817 201.87 -338.551 201.574 -338.278 201.284 -338 201 -336.489 199.523 -334.955 198.09 -333.401 196.698 -331.828 195.344 -330.238 194.025 -328.631 192.738 -325.375 190.25 -322.072 187.855 -318.734 185.531 -312 181 -298.641 173.094 -285.375 165.75 -271.922 158.781 -258 152 -256.79 151.52 -255.597 151.016 -254.422 150.49 -253.264 149.941 -252.122 149.371 -250.997 148.779 -249.889 148.165 -248.797 147.531 -247.721 146.877 -246.661 146.202 -245.617 145.507 -244.588 144.793 -243.575 144.06 -242.576 143.308 -241.593 142.538 -240.625 141.75 -239.671 140.944 -238.732 140.122 -237.807 139.282 -236.896 138.426 -235.116 136.666 -233.391 134.844 -231.718 132.963 -230.096 131.027 -228.524 129.039 -227 127 -226.085 126.062 -225.213 125.123 -224.379 124.181 -223.578 123.234 -222.807 122.282 -222.061 121.322 -221.335 120.354 -220.625 119.375 -219.236 117.381 -217.859 115.328 -215 111 -214.815 110.728 -214.636 110.473 -214.463 110.235 -214.295 110.016 -214.132 109.813 -213.974 109.629 -213.897 109.543 -213.821 109.462 -213.746 109.385 -213.672 109.312 -213.599 109.244 -213.528 109.181 -213.457 109.121 -213.387 109.066 -213.353 109.041 -213.319 109.016 -213.285 108.992 -213.251 108.97 -213.218 108.948 -213.185 108.928 -213.152 108.909 -213.119 108.891 -213.087 108.874 -213.054 108.858 -213.022 108.843 -212.991 108.829 -212.959 108.816 -212.928 108.805 -212.896 108.794 -212.865 108.785 -212.835 108.777 -212.804 108.77 -212.774 108.764 -212.744 108.759 -212.729 108.757 -212.714 108.755 -212.699 108.753 -212.684 108.752 -212.669 108.751 -212.654 108.751 -212.64 108.75 -212.625 108.75 -212.61 108.75 -212.596 108.751 -212.581 108.751 -212.567 108.752 -212.552 108.753 -212.538 108.755 -212.524 108.757 -212.509 108.759 -212.495 108.761 -212.481 108.764 -212.467 108.767 -212.452 108.77 -212.438 108.773 -212.424 108.777 -212.41 108.781 -212.396 108.785 -212.382 108.79 -212.368 108.794 -212.355 108.8 -212.341 108.805 -212.327 108.811 -212.313 108.816 -212.299 108.823 -212.286 108.829 -212.259 108.843 -212.231 108.858 -212.205 108.874 -212.178 108.891 -212.151 108.909 -212.125 108.928 -212.098 108.948 -212.072 108.97 -212.046 108.992 -212.02 109.016 -211.994 109.041 -211.968 109.066 -211.942 109.093 -211.917 109.121 -211.891 109.15 -211.866 109.181 -211.841 109.212 -211.816 109.244 -211.791 109.278 -211.766 109.312 -211.716 109.385 -211.667 109.462 -211.618 109.543 -211.569 109.629 -211.521 109.719 -211.473 109.813 -211.425 109.912 -211.377 110.016 -211.282 110.235 -211.188 110.473 -211.094 110.728 -211 111 -209.724 113.637 -208.525 116.297 -207.412 118.98 -206.889 120.331 -206.391 121.688 -205.917 123.05 -205.469 124.418 -205.048 125.792 -204.654 127.172 -204.289 128.558 -203.954 129.949 -203.649 131.347 -203.375 132.75 -203.133 134.159 -202.925 135.574 -202.751 136.995 -202.611 138.422 -202.508 139.854 -202.441 141.293 -202.412 142.737 -202.422 144.188 -202.471 145.644 -202.561 147.105 -202.692 148.573 -202.865 150.047 -203.082 151.526 -203.343 153.012 -203.648 154.503 -204 156 -205.41 162.375 -206.656 168.75 -207.762 175.125 -208.75 181.5 -210.469 194.25 -212 207 -212.105 208.312 -212.234 209.625 -212.561 212.246 -212.978 214.862 -213.484 217.469 -214.079 220.064 -214.76 222.645 -215.526 225.208 -216.375 227.75 -217.306 230.269 -218.318 232.762 -219.409 235.225 -220.578 237.656 -221.823 240.052 -223.143 242.41 -224.535 244.727 -226 247 -226.048 247.14 -226.1 247.278 -226.153 247.415 -226.21 247.551 -226.27 247.685 -226.332 247.818 -226.396 247.95 -226.463 248.08 -226.532 248.209 -226.604 248.337 -226.678 248.463 -226.754 248.589 -226.912 248.836 -227.078 249.078 -227.252 249.316 -227.432 249.55 -227.618 249.78 -227.811 250.006 -228.008 250.228 -228.21 250.447 -228.625 250.875 -229.052 251.291 -229.486 251.697 -230.359 252.484 -231.209 253.248 -231.614 253.625 -232 254 -241.426 260.607 -250.969 266.922 -260.652 272.932 -270.5 278.625 -275.493 281.349 -280.535 283.99 -285.63 286.546 -290.781 289.016 -295.991 291.397 -301.262 293.689 -306.597 295.891 -312 298 -313.502 298.744 -314.257 299.105 -315.016 299.453 -315.781 299.783 -316.553 300.092 -316.942 300.236 -317.334 300.374 -317.728 300.504 -318.125 300.625 -318.525 300.738 -318.928 300.841 -319.334 300.935 -319.744 301.018 -320.158 301.09 -320.575 301.15 -320.996 301.199 -321.422 301.234 -321.852 301.257 -322.286 301.266 -322.726 301.26 -323.17 301.24 -323.619 301.205 -324.074 301.153 -324.534 301.085 -325 301 m -415 82 l -415.21 83.5 -415.463 84.998 -415.754 86.493 -416.078 87.984 -416.432 89.469 -416.811 90.947 -417.625 93.875 -418.486 96.756 -419.359 99.578 -420.209 102.33 -421 105 -421.364 105.938 -421.707 106.875 -422.344 108.75 -422.934 110.625 -423.5 112.5 -424.066 114.375 -424.656 116.25 -425.293 118.125 -425.636 119.062 -426 120 -426.182 120.469 -426.355 120.937 -426.517 121.406 -426.669 121.875 -426.812 122.343 -426.945 122.812 -427.069 123.28 -427.184 123.748 -427.289 124.216 -427.385 124.684 -427.473 125.151 -427.551 125.618 -427.621 126.085 -427.683 126.552 -427.736 127.018 -427.781 127.484 -427.847 128.415 -427.882 129.344 -427.887 130.272 -427.863 131.197 -427.811 132.12 -427.733 133.041 -427.629 133.96 -427.5 134.875 -427.348 135.788 -427.173 136.697 -426.978 137.603 -426.762 138.506 -426.527 139.405 -426.274 140.3 -426.004 141.191 -425.719 142.078 -425.105 143.839 -424.441 145.58 -423.737 147.301 -423 149 -422.332 150.21 -421.641 151.403 -420.928 152.579 -420.193 153.738 -419.437 154.882 -418.66 156.009 -417.863 157.122 -417.047 158.219 -416.211 159.301 -415.357 160.37 -413.596 162.465 -411.767 164.507 -409.875 166.5 -407.925 168.446 -405.92 170.348 -403.865 172.208 -401.766 174.031 -399.625 175.819 -397.447 177.574 -393 181 -382.875 187.5 -377.391 190.844 -372 194 -371.396 194.288 -370.805 194.589 -370.225 194.902 -369.657 195.228 -369.101 195.566 -368.557 195.916 -368.025 196.277 -367.506 196.65 -366.999 197.035 -366.504 197.43 -366.021 197.836 -365.551 198.252 -365.094 198.678 -364.649 199.114 -364.216 199.56 -363.797 200.016 -363.39 200.48 -362.996 200.954 -362.616 201.436 -362.248 201.926 -361.893 202.425 -361.552 202.931 -361.223 203.445 -360.908 203.967 -360.607 204.495 -360.318 205.031 -360.044 205.573 -359.782 206.122 -359.535 206.677 -359.301 207.237 -359.081 207.803 -358.875 208.375 -358.683 208.952 -358.504 209.533 -358.34 210.12 -358.19 210.71 -358.054 211.305 -357.933 211.904 -357.825 212.506 -357.732 213.111 -357.654 213.72 -357.59 214.332 -357.541 214.946 -357.506 215.562 -357.486 216.181 -357.481 216.801 -357.491 217.423 -357.516 218.047 -357.555 218.671 -357.61 219.297 -357.68 219.923 -357.765 220.549 -357.866 221.175 -357.982 221.802 -358.113 222.428 -358.26 223.053 -358.422 223.677 -358.6 224.3 -358.794 224.922 -359.003 225.542 -359.228 226.16 -359.47 226.776 -359.727 227.389 -360 228 -363.703 234.75 -365.467 238.125 -367.125 241.5 -367.904 243.188 -368.643 244.875 -369.338 246.562 -369.984 248.25 -370.578 249.938 -371.115 251.625 -371.361 252.469 -371.591 253.312 -371.804 254.156 -372 255 -372.147 255.695 -372.307 256.374 -372.48 257.038 -372.666 257.686 -372.863 258.319 -373.073 258.937 -373.294 259.54 -373.527 260.129 -373.772 260.703 -374.027 261.264 -374.294 261.81 -374.571 262.343 -374.858 262.863 -375.156 263.369 -375.464 263.863 -375.781 264.344 -376.108 264.812 -376.445 265.268 -376.79 265.713 -377.144 266.145 -377.507 266.566 -377.878 266.976 -378.257 267.374 -378.645 267.762 -379.039 268.139 -379.442 268.505 -379.851 268.862 -380.268 269.208 -380.691 269.545 -381.121 269.873 -381.558 270.191 -382 270.5 -382.448 270.8 -382.902 271.092 -383.361 271.376 -383.826 271.651 -384.769 272.178 -385.73 272.676 -386.708 273.146 -387.7 273.589 -388.704 274.009 -389.719 274.406 -390.742 274.783 -391.773 275.141 -393.848 275.809 -395.928 276.423 -398 277 -398.841 277.176 -399.676 277.328 -400.505 277.457 -401.328 277.562 -402.146 277.645 -402.957 277.703 -403.763 277.738 -404.562 277.75 -405.356 277.738 -406.145 277.703 -406.927 277.645 -407.703 277.562 -408.474 277.457 -409.238 277.328 -409.997 277.176 -410.75 277 -411.497 276.801 -412.238 276.578 -412.974 276.332 -413.703 276.062 -414.427 275.77 -415.145 275.453 -415.856 275.113 -416.562 274.75 -417.263 274.363 -417.957 273.953 -418.646 273.52 -419.328 273.062 -420.005 272.582 -420.676 272.078 -421.341 271.551 -422 271 -423.852 269.477 -425.66 267.912 -427.427 266.309 -429.156 264.672 -430.85 263.006 -432.512 261.314 -435.75 257.875 -438.895 254.389 -441.969 250.891 -448 244 -449.137 242.7 -450.299 241.428 -451.487 240.188 -452.703 238.984 -453.948 237.822 -455.225 236.705 -455.875 236.165 -456.533 235.638 -457.2 235.124 -457.875 234.625 -458.559 234.14 -459.252 233.671 -459.954 233.217 -460.666 232.779 -461.387 232.359 -462.118 231.955 -462.859 231.57 -463.609 231.203 -464.37 230.855 -465.142 230.527 -465.924 230.219 -466.717 229.932 -467.521 229.665 -468.336 229.421 -469.162 229.199 -470 229 -471.114 228.801 -472.205 228.58 -473.276 228.339 -474.328 228.078 -475.363 227.8 -476.381 227.506 -478.375 226.875 -480.322 226.197 -482.234 225.484 -486 224 -486.363 223.823 -486.701 223.664 -487.013 223.517 -487.158 223.445 -487.297 223.375 -487.428 223.305 -487.491 223.269 -487.552 223.233 -487.61 223.197 -487.667 223.161 -487.722 223.124 -487.775 223.086 -487.826 223.047 -487.875 223.008 -487.922 222.968 -487.945 222.947 -487.967 222.927 -487.989 222.906 -488.01 222.884 -488.03 222.863 -488.05 222.841 -488.07 222.819 -488.089 222.796 -488.107 222.773 -488.125 222.75 -488.142 222.726 -488.159 222.702 -488.175 222.678 -488.191 222.653 -488.206 222.628 -488.22 222.602 -488.234 222.576 -488.248 222.55 -488.261 222.523 -488.273 222.495 -488.284 222.467 -488.296 222.439 -488.306 222.41 -488.316 222.381 -488.325 222.351 -488.334 222.32 -488.342 222.289 -488.35 222.258 -488.357 222.226 -488.363 222.193 -488.369 222.159 -488.374 222.125 -488.378 222.091 -488.382 222.056 -488.385 222.02 -488.388 221.983 -488.39 221.946 -488.391 221.908 -488.392 221.831 -488.391 221.75 -488.386 221.666 -488.38 221.58 -488.37 221.49 -488.358 221.397 -488.343 221.301 -488.326 221.202 -488.306 221.099 -488.283 220.992 -488.229 220.768 -488.164 220.528 -488.088 220.273 -488 220 -487.801 218.875 -487.582 217.75 -487.094 215.5 -486.559 213.25 -486 211 -485.441 208.75 -484.906 206.5 -484.418 204.25 -484.199 203.125 -484 202 -483.836 200.511 -483.719 199.045 -483.648 197.599 -483.625 196.172 -483.648 194.762 -483.719 193.369 -483.836 191.99 -483.912 191.306 -484 190.625 -484.1 189.947 -484.211 189.271 -484.334 188.598 -484.469 187.928 -484.773 186.593 -485.125 185.266 -485.523 183.944 -485.969 182.627 -486.461 181.313 -487 180 -487.724 178.5 -488.396 177 -489.016 175.5 -489.586 174 -490.106 172.5 -490.577 171 -491 169.5 -491.375 168 -491.703 166.5 -491.985 165 -492.222 163.5 -492.414 162 -492.562 160.5 -492.667 159 -492.729 157.5 -492.75 156 -492.73 154.5 -492.669 153 -492.569 151.5 -492.43 150 -492.253 148.5 -492.038 147 -491.787 145.5 -491.5 144 -491.178 142.5 -490.821 141 -490.431 139.5 -490.008 138 -489.552 136.5 -489.065 135 -488.548 133.5 -488 132 -485.326 124.922 -482.547 117.938 -479.65 111.047 -476.625 104.25 -473.459 97.547 -470.141 90.938 -466.658 84.422 -463 78 -462.426 77.063 -461.83 76.127 -460.578 74.266 -459.256 72.428 -457.875 70.625 -456.447 68.869 -454.984 67.172 -453.498 65.545 -452 64 -451.625 63.725 -451.25 63.461 -450.874 63.209 -450.498 62.969 -450.121 62.74 -449.743 62.523 -449.365 62.318 -448.984 62.125 -448.603 61.943 -448.219 61.773 -447.834 61.615 -447.447 61.469 -447.058 61.334 -446.666 61.211 -446.272 61.1 -445.875 61 -445.475 60.912 -445.072 60.836 -444.666 60.771 -444.256 60.719 -443.842 60.678 -443.425 60.648 -443.004 60.631 -442.578 60.625 -442.148 60.631 -441.714 60.648 -441.274 60.678 -440.83 60.719 -440.381 60.771 -439.926 60.836 -439.466 60.912 -439 61 -437.512 61.211 -436.047 61.469 -434.605 61.773 -433.894 61.943 -433.188 62.125 -432.487 62.318 -431.793 62.523 -431.104 62.74 -430.422 62.969 -429.745 63.209 -429.074 63.461 -428.409 63.725 -427.75 64 -427.097 64.287 -426.449 64.586 -425.808 64.896 -425.172 65.219 -424.542 65.553 -423.918 65.898 -423.3 66.256 -422.688 66.625 -422.081 67.006 -421.48 67.398 -420.886 67.803 -420.297 68.219 -419.714 68.646 -419.137 69.086 -418.565 69.537 -418 70 -417.668 70.334 -417.53 70.477 -417.406 70.609 -417.294 70.736 -417.242 70.799 -417.191 70.861 -417.142 70.925 -417.094 70.99 -417.047 71.056 -417 71.125 -416.953 71.197 -416.906 71.272 -416.809 71.436 -416.706 71.619 -416.594 71.828 -416.332 72.338 -416 73 m -158 304 l -158.222 306.801 -158.514 309.58 -158.872 312.339 -159.297 315.078 -159.786 317.8 -160.338 320.506 -160.951 323.197 -161.625 325.875 -162.357 328.541 -163.146 331.197 -163.991 333.844 -164.891 336.484 -165.843 339.118 -166.846 341.748 -169 347 -169.048 347.14 -169.1 347.278 -169.154 347.415 -169.21 347.551 -169.27 347.685 -169.332 347.817 -169.397 347.948 -169.465 348.078 -169.535 348.206 -169.608 348.333 -169.683 348.458 -169.76 348.582 -169.84 348.704 -169.922 348.825 -170.007 348.945 -170.094 349.062 -170.183 349.179 -170.274 349.294 -170.462 349.52 -170.659 349.739 -170.863 349.953 -171.075 350.161 -171.293 350.363 -171.519 350.56 -171.75 350.75 -171.987 350.935 -172.23 351.113 -172.478 351.286 -172.73 351.453 -172.987 351.614 -173.249 351.77 -173.513 351.919 -173.781 352.062 -174.052 352.2 -174.326 352.332 -174.601 352.458 -174.879 352.578 -175.158 352.692 -175.438 352.801 -175.719 352.903 -176 353 -176.753 353.457 -177.511 353.891 -178.276 354.302 -179.045 354.691 -179.819 355.058 -180.599 355.404 -181.383 355.728 -182.172 356.031 -182.965 356.314 -183.762 356.577 -184.564 356.82 -185.369 357.043 -186.178 357.247 -186.99 357.433 -187.806 357.601 -188.625 357.75 -189.447 357.882 -190.271 357.997 -191.098 358.094 -191.928 358.176 -192.759 358.241 -193.593 358.291 -194.428 358.325 -195.266 358.344 -196.104 358.348 -196.944 358.338 -198.627 358.277 -200.313 358.164 -202 358 -202.275 358.003 -202.539 358.012 -202.792 358.026 -203.033 358.047 -203.264 358.073 -203.483 358.105 -203.589 358.124 -203.692 358.144 -203.793 358.165 -203.891 358.188 -203.986 358.212 -204.079 358.237 -204.169 358.264 -204.257 358.293 -204.342 358.323 -204.425 358.354 -204.506 358.387 -204.584 358.422 -204.66 358.458 -204.733 358.495 -204.804 358.534 -204.873 358.574 -204.939 358.616 -205.003 358.659 -205.065 358.704 -205.125 358.75 -205.182 358.798 -205.238 358.847 -205.291 358.897 -205.342 358.949 -205.367 358.976 -205.391 359.003 -205.415 359.03 -205.438 359.058 -205.46 359.086 -205.483 359.114 -205.525 359.172 -205.566 359.231 -205.605 359.292 -205.642 359.354 -205.677 359.418 -205.709 359.483 -205.74 359.55 -205.77 359.618 -205.797 359.688 -205.822 359.759 -205.846 359.831 -205.868 359.905 -205.888 359.98 -205.906 360.057 -205.923 360.136 -205.938 360.216 -205.951 360.297 -205.963 360.38 -205.973 360.464 -205.981 360.55 -205.988 360.637 -205.993 360.725 -205.997 360.815 -206 361 -207.5 363.984 -209 366.875 -209.75 368.256 -210.5 369.578 -211.25 370.83 -211.625 371.426 -212 372 -212.563 372.645 -213.126 373.265 -213.69 373.862 -214.256 374.436 -214.824 374.984 -215.395 375.509 -215.969 376.009 -216.547 376.484 -217.129 376.935 -217.717 377.36 -218.309 377.76 -218.608 377.951 -218.908 378.135 -219.21 378.312 -219.514 378.484 -219.819 378.649 -220.126 378.807 -220.435 378.959 -220.746 379.104 -221.06 379.243 -221.375 379.375 -221.693 379.501 -222.012 379.62 -222.334 379.732 -222.659 379.838 -222.986 379.937 -223.315 380.029 -223.648 380.115 -223.982 380.193 -224.32 380.265 -224.66 380.331 -225.004 380.389 -225.35 380.441 -225.699 380.485 -226.051 380.523 -226.407 380.554 -226.766 380.578 -227.128 380.595 -227.493 380.605 -227.862 380.608 -228.234 380.604 -228.61 380.593 -228.99 380.575 -229.373 380.55 -229.76 380.518 -230.545 380.431 -231.347 380.317 -232.165 380.173 -233 380 -235.25 379.414 -237.498 378.783 -239.743 378.108 -241.984 377.391 -246.447 375.834 -250.875 374.125 -255.256 372.275 -259.578 370.297 -263.83 368.201 -268 366 -268.841 365.546 -269.676 365.121 -270.507 364.726 -271.332 364.359 -272.153 364.022 -272.97 363.715 -273.784 363.437 -274.594 363.188 -275.401 362.968 -276.206 362.777 -277.008 362.616 -277.409 362.547 -277.809 362.484 -278.208 362.429 -278.608 362.382 -279.007 362.342 -279.406 362.309 -279.804 362.283 -280.203 362.265 -280.602 362.254 -281 362.25 -281.398 362.254 -281.797 362.265 -282.196 362.283 -282.594 362.309 -282.993 362.342 -283.392 362.382 -283.792 362.429 -284.191 362.484 -284.591 362.547 -284.992 362.616 -285.794 362.777 -286.599 362.968 -287.406 363.188 -288.216 363.437 -289.03 363.715 -289.847 364.022 -290.668 364.359 -291.493 364.726 -292.324 365.121 -293.159 365.546 -294 366 -295.699 366.914 -297.422 367.781 -299.168 368.602 -300.938 369.375 -302.73 370.102 -304.547 370.781 -306.387 371.414 -308.25 372 -310.137 372.539 -312.047 373.031 -313.98 373.477 -315.938 373.875 -317.918 374.227 -319.922 374.531 -321.949 374.789 -324 375 -324.844 375.085 -325.687 375.152 -326.53 375.201 -327.371 375.23 -328.211 375.241 -329.049 375.233 -329.885 375.205 -330.719 375.156 -331.549 375.087 -332.376 374.998 -333.2 374.887 -334.02 374.754 -334.835 374.599 -335.645 374.422 -336.45 374.223 -337.25 374 -338.044 373.754 -338.832 373.484 -339.613 373.19 -340.387 372.871 -341.154 372.528 -341.913 372.159 -342.664 371.764 -343.406 371.344 -344.14 370.897 -344.865 370.423 -345.58 369.923 -346.285 369.395 -346.98 368.839 -347.665 368.254 -348.338 367.642 -349 367 -349.187 366.636 -349.369 366.293 -349.543 365.967 -349.703 365.656 -349.777 365.505 -349.846 365.357 -349.909 365.211 -349.939 365.138 -349.967 365.066 -349.993 364.995 -350.018 364.924 -350.04 364.853 -350.061 364.782 -350.08 364.711 -350.097 364.641 -350.112 364.57 -350.125 364.5 -350.136 364.43 -350.144 364.359 -350.15 364.289 -350.152 364.254 -350.154 364.218 -350.154 364.183 -350.155 364.147 -350.154 364.112 -350.153 364.076 -350.152 364.041 -350.149 364.005 -350.146 363.969 -350.143 363.934 -350.138 363.898 -350.133 363.862 -350.127 363.825 -350.121 363.789 -350.114 363.753 -350.106 363.716 -350.097 363.68 -350.088 363.643 -350.077 363.606 -350.067 363.569 -350.055 363.532 -350.042 363.495 -350.015 363.42 -349.984 363.344 -349.95 363.267 -349.913 363.19 -349.873 363.112 -349.828 363.033 -349.781 362.953 -349.729 362.872 -349.674 362.79 -349.615 362.707 -349.553 362.623 -349.486 362.538 -349.415 362.451 -349.341 362.364 -349.262 362.275 -349.179 362.185 -349 362 -348.089 360.775 -347.229 359.54 -346.422 358.293 -345.664 357.035 -344.956 355.767 -344.298 354.49 -343.688 353.203 -343.125 351.906 -342.609 350.601 -342.14 349.288 -341.715 347.966 -341.336 346.637 -341 345.3 -340.708 343.957 -340.458 342.606 -340.25 341.25 -340.083 339.888 -339.956 338.52 -339.869 337.147 -339.82 335.77 -339.81 334.388 -339.837 333.001 -339.901 331.612 -340 330.219 -340.135 328.823 -340.304 327.424 -340.506 326.024 -340.742 324.621 -341.01 323.217 -341.31 321.812 -341.64 320.406 -342 319 -342 316 -342.175 314.944 -342.318 314.021 -342.376 313.608 -342.424 313.225 -342.46 312.871 -342.474 312.706 -342.484 312.547 -342.491 312.395 -342.495 312.25 -342.495 312.112 -342.492 311.98 -342.485 311.855 -342.48 311.795 -342.474 311.736 -342.467 311.679 -342.459 311.624 -342.45 311.57 -342.439 311.518 -342.428 311.467 -342.416 311.417 -342.403 311.369 -342.388 311.323 -342.373 311.278 -342.356 311.234 -342.348 311.213 -342.338 311.192 -342.329 311.172 -342.32 311.152 -342.31 311.132 -342.299 311.112 -342.289 311.093 -342.278 311.074 -342.267 311.056 -342.256 311.038 -342.244 311.02 -342.232 311.003 -342.22 310.985 -342.207 310.969 -342.194 310.952 -342.181 310.936 -342.168 310.92 -342.154 310.905 -342.139 310.89 -342.125 310.875 -342.11 310.861 -342.095 310.846 -342.08 310.832 -342.064 310.819 -342.048 310.806 -342.031 310.793 -342.015 310.78 -341.997 310.768 -341.98 310.756 -341.962 310.744 -341.944 310.733 -341.926 310.722 -341.907 310.711 -341.888 310.701 -341.868 310.69 -341.848 310.68 -341.828 310.671 -341.808 310.662 -341.787 310.652 -341.766 310.644 -341.722 310.627 -341.677 310.612 -341.631 310.597 -341.583 310.584 -341.533 310.572 -341.482 310.561 -341.43 310.55 -341.376 310.541 -341.321 310.533 -341.264 310.526 -341.205 310.52 -341.145 310.515 -341.02 310.508 -340.888 310.505 -340.75 310.505 -340.605 310.509 -340.453 310.516 -340.294 310.526 -340.129 310.54 -339.775 310.576 -339.392 310.624 -338.979 310.682 -338.056 310.825 -337 311 -333.25 311.734 -331.375 312.072 -329.5 312.375 -327.625 312.631 -326.688 312.738 -325.75 312.828 -324.812 312.901 -323.875 312.955 -322.938 312.989 -322 313 -320.69 313.167 -319.387 313.295 -318.089 313.384 -316.797 313.436 -315.511 313.452 -314.23 313.433 -312.956 313.381 -311.688 313.297 -310.425 313.182 -309.168 313.038 -307.917 312.865 -306.672 312.666 -305.433 312.441 -304.199 312.192 -302.972 311.919 -301.75 311.625 -299.324 310.976 -296.922 310.256 -294.543 309.474 -292.188 308.641 -289.855 307.766 -287.547 306.861 -283 305 -279.273 303.465 -275.594 301.857 -271.961 300.177 -268.375 298.422 -264.836 296.591 -261.344 294.682 -257.898 292.694 -254.5 290.625 -251.148 288.474 -247.844 286.24 -244.586 283.921 -241.375 281.516 -238.211 279.022 -235.094 276.439 -232.023 273.766 -229 271 -226.962 269.112 -224.975 267.195 -223.043 265.243 -221.172 263.25 -219.365 261.21 -217.627 259.117 -216.785 258.049 -215.962 256.966 -215.159 255.866 -214.375 254.75 -213.612 253.616 -212.87 252.464 -212.149 251.293 -211.451 250.102 -210.776 248.89 -210.123 247.657 -209.495 246.402 -208.891 245.125 -208.311 243.824 -207.758 242.499 -207.23 241.149 -206.729 239.773 -206.254 238.371 -205.808 236.942 -205.39 235.485 -205 234 -204.46 235.687 -203.959 237.371 -203.047 240.719 -202.205 244.02 -201.375 247.25 -200.946 248.832 -200.498 250.387 -200.024 251.913 -199.516 253.406 -199.247 254.14 -198.967 254.865 -198.674 255.58 -198.369 256.285 -198.05 256.98 -197.716 257.665 -197.366 258.338 -197 259 -196.613 260.102 -196.41 260.636 -196.201 261.158 -195.985 261.67 -195.763 262.171 -195.533 262.661 -195.297 263.141 -195.053 263.61 -194.802 264.07 -194.542 264.519 -194.275 264.959 -194 265.389 -193.717 265.81 -193.425 266.222 -193.125 266.625 -192.816 267.019 -192.498 267.405 -192.171 267.782 -191.834 268.15 -191.488 268.511 -191.132 268.864 -190.766 269.209 -190.391 269.547 -190.005 269.877 -189.608 270.2 -189.201 270.517 -188.783 270.826 -188.354 271.129 -187.914 271.426 -187.463 271.716 -187 272 -183.998 273.545 -180.984 275.172 -174.875 278.625 -168.578 282.266 -162 286 -161.725 286.188 -161.461 286.375 -161.208 286.564 -160.967 286.754 -160.736 286.945 -160.517 287.138 -160.308 287.333 -160.109 287.531 -160.014 287.631 -159.921 287.732 -159.831 287.834 -159.743 287.936 -159.658 288.039 -159.575 288.144 -159.494 288.249 -159.416 288.355 -159.34 288.463 -159.267 288.572 -159.196 288.681 -159.127 288.792 -159.061 288.905 -158.997 289.018 -158.935 289.134 -158.875 289.25 -158.818 289.368 -158.762 289.487 -158.709 289.608 -158.658 289.731 -158.609 289.855 -158.562 289.981 -158.517 290.109 -158.475 290.238 -158.434 290.37 -158.395 290.503 -158.323 290.775 -158.26 291.055 -158.203 291.344 -158.154 291.641 -158.112 291.948 -158.077 292.264 -158.049 292.59 -158.027 292.926 -158.012 293.273 -158.003 293.631 -158 294 -158 304 m -197 109 l -192.875 109.223 -188.75 109.518 -184.625 109.885 -180.5 110.328 -176.375 110.847 -172.25 111.443 -168.125 112.119 -164 112.875 -159.875 113.713 -155.75 114.635 -151.625 115.641 -147.5 116.734 -143.375 117.915 -139.25 119.186 -135.125 120.547 -131 122 -129.605 122.48 -128.234 122.985 -126.887 123.513 -125.562 124.064 -124.262 124.641 -122.984 125.241 -121.73 125.866 -120.5 126.516 -119.293 127.19 -118.109 127.89 -116.949 128.615 -115.812 129.365 -114.699 130.141 -113.609 130.943 -112.543 131.771 -111.5 132.625 -110.48 133.505 -109.484 134.412 -108.512 135.346 -107.562 136.307 -106.637 137.294 -105.734 138.309 -104.855 139.352 -104 140.422 -103.168 141.52 -102.359 142.646 -101.574 143.8 -100.812 144.982 -100.074 146.194 -99.359 147.433 -98.668 148.702 -98 150 -95.328 155.58 -92.562 161.078 -89.703 166.506 -86.75 171.875 -80.562 182.484 -74 193 -73.623 193.381 -73.431 193.582 -73.234 193.797 -73.032 194.029 -72.822 194.283 -72.604 194.564 -72.375 194.875 -72.135 195.221 -71.881 195.607 -71.613 196.037 -71.328 196.516 -71.026 197.047 -70.705 197.635 -70 199 -69.813 199.375 -69.633 199.748 -69.547 199.934 -69.464 200.118 -69.386 200.302 -69.312 200.484 -69.245 200.665 -69.185 200.844 -69.157 200.933 -69.131 201.022 -69.108 201.11 -69.086 201.197 -69.067 201.284 -69.049 201.37 -69.035 201.456 -69.022 201.541 -69.013 201.626 -69.006 201.71 -69.001 201.793 -69 201.834 -69 201.875 -69 201.916 -69.001 201.957 -69.003 201.997 -69.006 202.038 -69.009 202.078 -69.014 202.118 -69.019 202.157 -69.024 202.197 -69.031 202.236 -69.039 202.276 -69.047 202.314 -69.056 202.353 -69.066 202.392 -69.077 202.43 -69.089 202.468 -69.102 202.506 -69.115 202.543 -69.13 202.581 -69.145 202.618 -69.162 202.655 -69.179 202.692 -69.198 202.728 -69.217 202.764 -69.237 202.8 -69.259 202.836 -69.281 202.871 -69.305 202.906 -69.329 202.941 -69.355 202.976 -69.381 203.01 -69.409 203.044 -69.438 203.078 -69.467 203.112 -69.498 203.145 -69.53 203.178 -69.564 203.211 -69.598 203.243 -69.633 203.275 -69.67 203.307 -69.708 203.339 -69.747 203.37 -69.787 203.401 -69.829 203.431 -69.871 203.462 -69.915 203.492 -69.961 203.522 -70.055 203.58 -70.154 203.637 -70.258 203.693 -70.368 203.748 -70.483 203.801 -70.604 203.853 -70.73 203.903 -70.862 203.952 -71 204 -78.312 207.594 -85.25 211 -88.578 212.691 -91.812 214.406 -94.953 216.168 -96.488 217.074 -98 218 -98.738 218.563 -99.453 219.127 -100.145 219.694 -100.812 220.266 -101.457 220.843 -102.078 221.428 -102.676 222.021 -103.25 222.625 -103.801 223.24 -104.328 223.869 -104.832 224.512 -105.075 224.84 -105.312 225.172 -105.544 225.508 -105.77 225.849 -105.989 226.194 -106.203 226.545 -106.411 226.901 -106.613 227.261 -106.81 227.628 -107 228 -109.678 233.197 -111.063 235.75 -112.484 238.266 -113.947 240.74 -115.455 243.17 -117.013 245.55 -118.625 247.875 -120.296 250.142 -122.029 252.346 -122.921 253.423 -123.83 254.482 -124.757 255.524 -125.703 256.547 -126.668 257.551 -127.652 258.535 -128.657 259.5 -129.682 260.443 -130.728 261.366 -131.796 262.266 -132.887 263.145 -134 264 -134.364 264.188 -134.707 264.375 -135.033 264.562 -135.344 264.75 -135.643 264.938 -135.934 265.125 -136.5 265.5 -137.066 265.875 -137.357 266.062 -137.656 266.25 -137.967 266.438 -138.293 266.625 -138.636 266.812 -139 267 -139.844 267.548 -140.688 268.066 -141.532 268.554 -142.377 269.012 -143.223 269.439 -144.069 269.834 -144.917 270.199 -145.766 270.531 -146.616 270.831 -147.042 270.969 -147.468 271.099 -147.895 271.221 -148.322 271.334 -148.75 271.439 -149.178 271.535 -149.606 271.623 -150.036 271.703 -150.466 271.774 -150.896 271.836 -151.327 271.89 -151.759 271.936 -152.192 271.972 -152.625 272 -153.059 272.019 -153.494 272.029 -153.929 272.031 -154.365 272.023 -154.803 272.006 -155.241 271.981 -155.679 271.946 -156.119 271.902 -156.56 271.849 -157.001 271.787 -157.444 271.716 -157.887 271.635 -158.332 271.545 -158.778 271.446 -159.224 271.337 -159.672 271.219 -160.121 271.091 -160.571 270.954 -161.022 270.807 -161.474 270.65 -161.927 270.484 -162.382 270.307 -162.838 270.122 -163.295 269.926 -164.213 269.505 -165.136 269.043 -166.065 268.542 -167 268 -173.566 263.334 -176.207 261.358 -177.383 260.435 -178.469 259.547 -179.47 258.687 -180.391 257.847 -180.823 257.434 -181.237 257.022 -181.633 256.613 -182.012 256.205 -182.374 255.797 -182.721 255.388 -183.052 254.978 -183.369 254.566 -183.671 254.15 -183.96 253.73 -184.236 253.305 -184.5 252.875 -184.752 252.438 -184.993 251.993 -185.223 251.541 -185.444 251.079 -185.655 250.607 -185.857 250.124 -186.052 249.63 -186.238 249.123 -186.418 248.603 -186.591 248.068 -186.921 246.953 -187.233 245.771 -187.531 244.516 -187.821 243.179 -188.106 241.756 -188.684 238.619 -190 231 -190.54 228.75 -191.037 226.498 -191.922 221.984 -192.689 217.447 -193.375 212.875 -194.641 203.578 -196 194 -196.085 193.06 -196.153 192.114 -196.205 191.163 -196.24 190.207 -196.266 188.283 -196.234 186.344 -196.15 184.393 -196.018 182.434 -195.841 180.468 -195.625 178.5 -195.374 176.532 -195.092 174.566 -194.453 170.656 -193.744 166.793 -193 163 -192.555 161.222 -192.155 159.449 -191.801 157.683 -191.49 155.922 -191.223 154.167 -190.998 152.418 -190.815 150.675 -190.672 148.938 -190.568 147.206 -190.504 145.48 -190.477 143.761 -190.486 142.047 -190.532 140.339 -190.613 138.637 -190.727 136.94 -190.875 135.25 -191.055 133.565 -191.266 131.887 -191.508 130.214 -191.779 128.547 -192.406 125.23 -193.141 121.938 -193.974 118.668 -194.9 115.422 -195.911 112.199 -197 109 m -349 192 l -348.579 188.637 -348.066 185.297 -347.466 181.98 -346.781 178.688 -346.014 175.418 -345.168 172.172 -344.246 168.949 -343.25 165.75 -342.184 162.574 -341.051 159.422 -339.853 156.293 -338.594 153.188 -337.276 150.105 -335.902 147.047 -333 141 -331.841 138.386 -330.619 135.795 -329.342 133.224 -328.016 130.672 -325.248 125.619 -322.375 120.625 -316.547 110.766 -313.709 105.877 -311 101 -309.898 99.301 -308.844 97.58 -307.836 95.839 -306.875 94.078 -305.961 92.3 -305.094 90.506 -304.273 88.697 -303.5 86.875 -302.773 85.041 -302.094 83.197 -301.461 81.344 -300.875 79.484 -300.336 77.618 -299.844 75.748 -299.398 73.875 -299 72 -298.824 71.162 -298.672 70.336 -298.543 69.521 -298.438 68.717 -298.355 67.924 -298.297 67.142 -298.262 66.37 -298.25 65.609 -298.262 64.859 -298.297 64.118 -298.355 63.387 -298.438 62.666 -298.543 61.954 -298.672 61.252 -298.745 60.905 -298.824 60.559 -298.909 60.216 -299 59.875 -299.097 59.536 -299.199 59.2 -299.308 58.865 -299.422 58.533 -299.542 58.203 -299.668 57.875 -299.8 57.549 -299.938 57.225 -300.081 56.903 -300.23 56.583 -300.547 55.948 -300.887 55.322 -301.25 54.703 -301.637 54.092 -302.047 53.487 -302.48 52.89 -302.938 52.299 -303.418 51.715 -303.922 51.137 -304.449 50.565 -305 50 -306.136 48.711 -307.295 47.467 -308.474 46.267 -309.672 45.109 -310.887 43.993 -312.119 42.916 -313.365 41.877 -314.625 40.875 -315.896 39.908 -317.178 38.975 -318.468 38.073 -319.766 37.203 -321.069 36.362 -322.377 35.549 -325 34 -325.551 33.625 -326.08 33.25 -326.589 32.875 -327.078 32.5 -327.55 32.125 -328.006 31.75 -328.875 31 -329.697 30.25 -330.484 29.5 -332 28 -332.093 27.952 -332.185 27.903 -332.275 27.853 -332.363 27.801 -332.45 27.748 -332.535 27.693 -332.619 27.637 -332.701 27.58 -332.782 27.522 -332.86 27.462 -332.938 27.401 -333.013 27.339 -333.087 27.275 -333.158 27.211 -333.229 27.145 -333.297 27.078 -333.363 27.01 -333.428 26.941 -333.491 26.871 -333.552 26.8 -333.61 26.728 -333.667 26.655 -333.722 26.581 -333.775 26.506 -333.826 26.43 -333.875 26.353 -333.922 26.276 -333.967 26.197 -334.01 26.118 -334.05 26.038 -334.089 25.957 -334.125 25.875 -334.159 25.793 -334.191 25.71 -334.22 25.626 -334.248 25.541 -334.273 25.456 -334.296 25.37 -334.316 25.284 -334.334 25.197 -334.35 25.11 -334.363 25.022 -334.374 24.933 -334.382 24.844 -334.388 24.755 -334.391 24.665 -334.392 24.575 -334.391 24.484 -334.386 24.393 -334.38 24.302 -334.37 24.21 -334.358 24.118 -334.343 24.026 -334.326 23.934 -334.306 23.841 -334.283 23.748 -334.258 23.655 -334.229 23.562 -334.198 23.468 -334.164 23.375 -334.128 23.281 -334.088 23.187 -334.045 23.094 -334 23 -333.999 22.908 -333.994 22.818 -333.991 22.775 -333.987 22.732 -333.982 22.689 -333.977 22.648 -333.971 22.607 -333.964 22.567 -333.957 22.527 -333.949 22.489 -333.94 22.45 -333.931 22.413 -333.921 22.376 -333.91 22.34 -333.899 22.304 -333.887 22.269 -333.874 22.235 -333.861 22.201 -333.847 22.168 -333.833 22.136 -333.818 22.104 -333.802 22.073 -333.786 22.042 -333.769 22.012 -333.752 21.983 -333.734 21.954 -333.715 21.925 -333.696 21.898 -333.676 21.87 -333.656 21.844 -333.635 21.818 -333.614 21.792 -333.592 21.767 -333.57 21.743 -333.547 21.719 -333.524 21.695 -333.5 21.672 -333.475 21.65 -333.45 21.628 -333.425 21.607 -333.399 21.586 -333.372 21.565 -333.345 21.545 -333.318 21.526 -333.29 21.507 -333.262 21.488 -333.233 21.47 -333.204 21.453 -333.174 21.435 -333.144 21.419 -333.082 21.386 -333.019 21.356 -332.954 21.327 -332.888 21.3 -332.82 21.274 -332.75 21.25 -332.679 21.227 -332.607 21.206 -332.533 21.186 -332.458 21.167 -332.381 21.15 -332.303 21.134 -332.145 21.105 -331.981 21.081 -331.814 21.061 -331.643 21.044 -331.469 21.031 -331.112 21.013 -330.746 21.004 -330 21 -328.5 20.836 -327 20.721 -325.5 20.655 -324.75 20.641 -324 20.641 -323.25 20.653 -322.5 20.679 -321.75 20.718 -321 20.771 -320.25 20.839 -319.5 20.92 -318.75 21.015 -318 21.125 -317.25 21.25 -316.5 21.389 -315.75 21.543 -315 21.713 -314.25 21.898 -313.5 22.098 -312.75 22.315 -312 22.547 -311.25 22.795 -310.5 23.06 -309.75 23.341 -309 23.639 -308.25 23.953 -307.5 24.285 -306.75 24.634 -306 25 -296.344 30.766 -286.875 36.812 -277.594 43.141 -268.5 49.75 -259.594 56.641 -250.875 63.812 -242.344 71.266 -234 79 -232.998 79.94 -232.055 80.887 -231.605 81.362 -231.17 81.839 -230.75 82.317 -230.344 82.797 -229.953 83.278 -229.576 83.761 -229.214 84.245 -228.867 84.73 -228.535 85.218 -228.217 85.706 -227.914 86.196 -227.625 86.688 -227.351 87.18 -227.092 87.675 -226.847 88.171 -226.617 88.668 -226.402 89.167 -226.201 89.667 -226.015 90.169 -225.844 90.672 -225.687 91.177 -225.545 91.683 -225.417 92.19 -225.305 92.699 -225.207 93.21 -225.123 93.722 -225.054 94.235 -225 94.75 -224.96 95.266 -224.936 95.784 -224.925 96.303 -224.93 96.824 -224.949 97.346 -224.982 97.87 -225.031 98.395 -225.094 98.922 -225.171 99.45 -225.264 99.979 -225.371 100.51 -225.492 101.043 -225.628 101.577 -225.779 102.112 -225.945 102.649 -226.125 103.188 -226.32 103.727 -226.529 104.269 -226.753 104.811 -226.992 105.355 -227.246 105.901 -227.514 106.448 -227.796 106.997 -228.094 107.547 -228.406 108.098 -228.732 108.651 -229.074 109.206 -229.43 109.762 -230.186 110.878 -231 112 -232.523 113.852 -234.092 115.658 -235.704 117.421 -237.359 119.141 -239.055 120.82 -240.791 122.459 -242.565 124.06 -244.375 125.625 -246.22 127.155 -248.1 128.65 -250.011 130.114 -251.953 131.547 -255.924 134.326 -260 137 -264.586 139.582 -269.312 142.094 -279 147 -288.688 151.906 -293.414 154.418 -298 157 -301.176 158.722 -304.328 160.512 -307.457 162.366 -310.562 164.281 -316.703 168.285 -322.75 172.5 -328.703 176.902 -334.562 181.469 -340.328 186.176 -346 191 -346.046 191.001 -346.091 191.003 -346.134 191.006 -346.177 191.011 -346.217 191.017 -346.257 191.025 -346.296 191.033 -346.334 191.043 -346.371 191.054 -346.407 191.066 -346.442 191.078 -346.477 191.092 -346.511 191.107 -346.544 191.123 -346.577 191.139 -346.609 191.156 -346.641 191.174 -346.673 191.193 -346.705 191.212 -346.736 191.232 -346.799 191.273 -346.861 191.316 -346.99 191.407 -347.056 191.453 -347.125 191.5 -347.197 191.547 -347.272 191.593 -347.352 191.639 -347.393 191.661 -347.436 191.684 -347.479 191.705 -347.525 191.727 -347.571 191.748 -347.619 191.768 -347.669 191.788 -347.72 191.807 -347.773 191.826 -347.828 191.844 -347.885 191.861 -347.943 191.877 -348.004 191.893 -348.066 191.908 -348.131 191.922 -348.198 191.934 -348.267 191.946 -348.338 191.957 -348.412 191.967 -348.488 191.975 -348.566 191.983 -348.648 191.989 -348.818 191.997 -349 192 m -163 442 l -167.832 443.783 -172.594 445.391 -177.309 446.834 -182 448.125 -186.691 449.275 -191.406 450.297 -196.168 451.201 -201 452 -204.199 452.703 -207.418 453.312 -210.655 453.828 -213.906 454.25 -217.169 454.578 -220.441 454.812 -223.719 454.953 -227 455 -230.281 454.953 -233.559 454.812 -236.831 454.578 -240.094 454.25 -243.345 453.828 -246.582 453.312 -249.801 452.703 -253 452 -255.593 451.309 -258.122 450.548 -260.587 449.719 -262.99 448.822 -265.331 447.859 -267.61 446.83 -269.827 445.736 -271.984 444.578 -274.081 443.357 -276.119 442.074 -278.097 440.73 -280.018 439.326 -281.88 437.863 -283.685 436.341 -285.433 434.761 -287.125 433.125 -288.761 431.433 -290.343 429.687 -291.869 427.886 -293.342 426.033 -294.761 424.128 -296.127 422.172 -297.441 420.165 -298.703 418.109 -299.914 416.005 -301.074 413.854 -302.184 411.656 -303.244 409.412 -304.255 407.124 -305.218 404.792 -306.133 402.417 -307 400 -307.182 399.719 -307.353 399.437 -307.513 399.155 -307.664 398.873 -307.806 398.59 -307.94 398.306 -308.067 398.021 -308.188 397.734 -308.302 397.447 -308.411 397.157 -308.617 396.572 -308.812 395.979 -309 395.375 -309.383 394.131 -309.589 393.488 -309.812 392.828 -310.06 392.151 -310.336 391.455 -310.487 391.099 -310.647 390.739 -310.818 390.372 -311 390 -311 389.802 -310.996 389.584 -310.992 389.469 -310.987 389.352 -310.979 389.232 -310.969 389.109 -310.956 388.986 -310.939 388.861 -310.919 388.736 -310.907 388.674 -310.895 388.611 -310.881 388.549 -310.866 388.487 -310.85 388.426 -310.833 388.365 -310.814 388.304 -310.794 388.244 -310.773 388.184 -310.75 388.125 -310.726 388.067 -310.7 388.009 -310.673 387.953 -310.644 387.897 -310.614 387.843 -310.581 387.789 -310.547 387.737 -310.53 387.711 -310.512 387.686 -310.493 387.66 -310.474 387.636 -310.455 387.611 -310.435 387.587 -310.414 387.563 -310.393 387.54 -310.372 387.517 -310.35 387.494 -310.328 387.472 -310.305 387.45 -310.281 387.429 -310.257 387.408 -310.233 387.387 -310.208 387.367 -310.182 387.347 -310.156 387.328 -310.13 387.309 -310.102 387.291 -310.075 387.273 -310.046 387.256 -310.017 387.239 -309.988 387.222 -309.958 387.207 -309.927 387.191 -309.896 387.176 -309.864 387.162 -309.832 387.148 -309.799 387.135 -309.765 387.122 -309.731 387.11 -309.696 387.099 -309.66 387.088 -309.624 387.078 -309.587 387.068 -309.55 387.059 -309.511 387.05 -309.473 387.042 -309.433 387.035 -309.393 387.029 -309.352 387.023 -309.311 387.017 -309.268 387.013 -309.182 387.006 -309.092 387.001 -309 387 -303.156 384.75 -297.5 382.5 -291.844 380.25 -286 378 -285.261 377.824 -284.545 377.674 -283.849 377.55 -283.508 377.498 -283.172 377.453 -282.84 377.416 -282.512 377.386 -282.189 377.364 -281.869 377.35 -281.553 377.343 -281.24 377.345 -280.931 377.356 -280.625 377.375 -280.322 377.403 -280.021 377.44 -279.723 377.486 -279.428 377.541 -279.134 377.606 -278.843 377.68 -278.553 377.765 -278.266 377.859 -277.979 377.964 -277.694 378.079 -277.41 378.205 -277.127 378.342 -276.845 378.489 -276.563 378.648 -276.281 378.818 -276 379 -274.125 379.772 -272.25 380.584 -268.5 382.297 -261 385.875 -257.25 387.623 -255.375 388.461 -253.5 389.266 -251.625 390.029 -249.75 390.744 -247.875 391.404 -246 392 -244.498 392.373 -242.984 392.734 -241.447 393.072 -239.875 393.375 -239.072 393.51 -238.256 393.631 -237.425 393.738 -236.578 393.828 -235.714 393.901 -234.83 393.955 -233.926 393.989 -233 394 -232.438 394.1 -231.875 394.21 -231.313 394.332 -230.752 394.465 -230.191 394.608 -229.632 394.76 -228.516 395.094 -227.406 395.462 -226.303 395.863 -225.209 396.293 -224.125 396.75 -223.053 397.23 -221.994 397.73 -220.95 398.249 -219.922 398.781 -217.92 399.879 -216 401 -212.46 403.108 -208.965 405.305 -205.51 407.583 -202.094 409.938 -198.712 412.362 -195.363 414.852 -188.75 420 -182.23 425.336 -175.781 430.812 -163 442 m -378 37 l -378.701 41.127 -379.297 45.266 -379.775 49.428 -380.125 53.625 -380.248 55.74 -380.334 57.869 -380.382 60.012 -380.391 62.172 -380.358 64.349 -380.283 66.545 -380.164 68.761 -380 71 -379.698 73.051 -379.354 75.077 -378.968 77.08 -378.539 79.057 -378.066 81.008 -377.548 82.933 -376.984 84.832 -376.375 86.703 -375.719 88.547 -375.015 90.362 -374.262 92.148 -373.461 93.904 -372.61 95.631 -371.708 97.327 -370.755 98.992 -369.75 100.625 -368.692 102.226 -367.581 103.794 -366.416 105.329 -365.195 106.83 -363.919 108.297 -362.587 109.728 -361.197 111.124 -359.75 112.484 -358.244 113.808 -356.679 115.094 -355.053 116.342 -353.367 117.553 -351.62 118.724 -349.81 119.856 -347.937 120.948 -346 122 -345.818 122.094 -345.648 122.188 -345.489 122.282 -345.413 122.329 -345.34 122.377 -345.269 122.425 -345.201 122.473 -345.136 122.521 -345.073 122.569 -345.012 122.618 -344.954 122.667 -344.898 122.716 -344.844 122.766 -344.792 122.816 -344.743 122.866 -344.695 122.917 -344.65 122.968 -344.607 123.02 -344.565 123.072 -344.526 123.125 -344.488 123.178 -344.453 123.231 -344.419 123.286 -344.386 123.341 -344.356 123.396 -344.327 123.452 -344.3 123.509 -344.274 123.567 -344.25 123.625 -344.227 123.684 -344.206 123.744 -344.186 123.804 -344.167 123.865 -344.15 123.928 -344.134 123.991 -344.119 124.054 -344.105 124.119 -344.093 124.185 -344.081 124.251 -344.061 124.387 -344.044 124.528 -344.031 124.672 -344.021 124.821 -344.013 124.974 -344.004 125.295 -344 125.636 -344 126 -346.58 132.418 -349.078 138.906 -353.875 152 -363 178 -363.177 178.363 -363.334 178.701 -363.477 179.013 -363.609 179.297 -363.673 179.428 -363.736 179.552 -363.799 179.667 -363.83 179.722 -363.861 179.775 -363.893 179.826 -363.925 179.875 -363.957 179.922 -363.99 179.967 -364.023 180.01 -364.039 180.03 -364.056 180.05 -364.073 180.07 -364.09 180.089 -364.108 180.107 -364.125 180.125 -364.143 180.142 -364.161 180.159 -364.179 180.175 -364.197 180.191 -364.215 180.206 -364.234 180.22 -364.253 180.234 -364.272 180.248 -364.292 180.261 -364.311 180.273 -364.331 180.284 -364.352 180.296 -364.372 180.306 -364.393 180.316 -364.414 180.325 -364.436 180.334 -364.457 180.342 -364.479 180.35 -364.502 180.357 -364.525 180.363 -364.548 180.369 -364.571 180.374 -364.595 180.378 -364.619 180.382 -364.644 180.385 -364.669 180.388 -364.695 180.39 -364.72 180.391 -364.747 180.392 -364.773 180.392 -364.801 180.392 -364.828 180.391 -364.856 180.389 -364.885 180.386 -364.914 180.383 -364.943 180.38 -364.973 180.375 -365.004 180.37 -365.035 180.365 -365.066 180.358 -365.131 180.343 -365.198 180.326 -365.267 180.306 -365.338 180.283 -365.412 180.258 -365.488 180.229 -365.566 180.198 -365.648 180.164 -365.818 180.088 -366 180 -368.602 178.664 -371.156 177.281 -373.664 175.852 -376.125 174.375 -378.539 172.852 -380.906 171.281 -383.227 169.664 -385.5 168 -387.727 166.289 -389.906 164.531 -392.039 162.727 -394.125 160.875 -396.164 158.977 -398.156 157.031 -400.102 155.039 -402 153 -402.925 151.875 -403.824 150.746 -404.694 149.612 -405.531 148.469 -406.333 147.314 -407.098 146.145 -407.821 144.958 -408.166 144.357 -408.5 143.75 -408.822 143.138 -409.132 142.519 -409.43 141.894 -409.715 141.262 -409.987 140.622 -410.245 139.975 -410.489 139.32 -410.719 138.656 -410.934 137.984 -411.134 137.302 -411.319 136.611 -411.488 135.91 -411.641 135.199 -411.778 134.477 -411.898 133.744 -412 133 -412.045 132.859 -412.088 132.719 -412.128 132.578 -412.165 132.438 -412.199 132.297 -412.231 132.157 -412.287 131.877 -412.334 131.598 -412.371 131.319 -412.401 131.042 -412.422 130.766 -412.436 130.491 -412.443 130.218 -412.444 129.947 -412.439 129.678 -412.43 129.411 -412.415 129.146 -412.375 128.625 -412.323 128.115 -412.264 127.619 -412.141 126.672 -412.086 126.224 -412.041 125.795 -412.024 125.588 -412.011 125.386 -412.003 125.19 -412 125 -410.83 120.5 -409.578 116 -406.875 107 -403.984 98 -401 89 -400.73 88.159 -400.484 87.324 -400.261 86.494 -400.061 85.67 -399.883 84.851 -399.728 84.036 -399.595 83.227 -399.484 82.422 -399.396 81.621 -399.329 80.825 -399.284 80.033 -399.26 79.244 -399.257 78.459 -399.276 77.678 -399.315 76.9 -399.375 76.125 -399.456 75.353 -399.556 74.584 -399.677 73.817 -399.818 73.053 -399.979 72.291 -400.159 71.531 -400.359 70.772 -400.578 70.016 -400.816 69.26 -401.073 68.507 -401.349 67.754 -401.643 67.002 -401.955 66.251 -402.285 65.5 -402.634 64.75 -403 64 -403.27 63.443 -403.516 62.898 -403.74 62.364 -403.941 61.84 -404.121 61.326 -404.279 60.823 -404.35 60.575 -404.415 60.329 -404.476 60.085 -404.531 59.844 -404.581 59.605 -404.627 59.368 -404.667 59.133 -404.702 58.9 -404.732 58.669 -404.757 58.44 -404.777 58.213 -404.793 57.988 -404.804 57.765 -404.81 57.544 -404.811 57.324 -404.808 57.106 -404.8 56.89 -404.788 56.675 -404.771 56.462 -404.75 56.25 -404.724 56.04 -404.694 55.831 -404.66 55.624 -404.622 55.417 -404.579 55.213 -404.532 55.009 -404.481 54.807 -404.426 54.605 -404.367 54.405 -404.304 54.206 -404.236 54.008 -404.166 53.811 -404.091 53.615 -404.012 53.419 -403.93 53.225 -403.844 53.031 -403.661 52.646 -403.463 52.263 -403.252 51.883 -403.027 51.504 -402.789 51.127 -402.539 50.75 -402.275 50.375 -402 50 -401.763 49.582 -401.52 49.171 -401.271 48.767 -401.017 48.37 -400.757 47.981 -400.492 47.598 -400.221 47.222 -399.945 46.854 -399.664 46.491 -399.379 46.136 -399.088 45.787 -398.792 45.445 -398.492 45.109 -398.186 44.78 -397.562 44.141 -396.921 43.526 -396.263 42.936 -395.588 42.369 -394.898 41.826 -394.194 41.306 -393.476 40.807 -392.744 40.331 -392 39.875 -391.244 39.44 -390.478 39.025 -389.701 38.629 -388.914 38.252 -388.119 37.893 -387.315 37.552 -386.505 37.229 -385.688 36.922 -384.864 36.631 -384.036 36.356 -383.203 36.095 -382.367 35.85 -380.687 35.399 -379 35 -378.908 35.001 -378.821 35.003 -378.738 35.006 -378.659 35.011 -378.584 35.018 -378.513 35.026 -378.446 35.035 -378.383 35.045 -378.323 35.057 -378.295 35.063 -378.267 35.069 -378.24 35.076 -378.214 35.084 -378.189 35.091 -378.165 35.099 -378.142 35.107 -378.119 35.115 -378.097 35.124 -378.076 35.133 -378.056 35.142 -378.037 35.152 -378.018 35.162 -378 35.172 -377.983 35.182 -377.966 35.193 -377.951 35.204 -377.935 35.215 -377.921 35.227 -377.914 35.232 -377.907 35.238 -377.901 35.244 -377.894 35.25 -377.888 35.256 -377.882 35.262 -377.876 35.269 -377.87 35.275 -377.864 35.281 -377.859 35.288 -377.854 35.294 -377.848 35.301 -377.843 35.307 -377.839 35.314 -377.834 35.321 -377.829 35.327 -377.825 35.334 -377.82 35.341 -377.816 35.348 -377.812 35.355 -377.808 35.362 -377.805 35.369 -377.801 35.376 -377.798 35.384 -377.794 35.391 -377.791 35.398 -377.788 35.406 -377.785 35.413 -377.782 35.421 -377.78 35.428 -377.775 35.443 -377.77 35.459 -377.766 35.475 -377.763 35.49 -377.76 35.507 -377.757 35.523 -377.755 35.539 -377.753 35.556 -377.752 35.573 -377.751 35.59 -377.75 35.608 -377.75 35.625 -377.75 35.643 -377.751 35.661 -377.752 35.679 -377.753 35.697 -377.756 35.734 -377.761 35.771 -377.766 35.809 -377.773 35.848 -377.781 35.888 -377.789 35.928 -377.808 36.009 -377.829 36.093 -377.875 36.266 -377.898 36.354 -377.921 36.444 -377.942 36.535 -377.961 36.627 -377.969 36.673 -377.977 36.72 -377.984 36.766 -377.989 36.813 -377.994 36.859 -377.997 36.906 -377.999 36.953 -378 37 m -653 401 l -624 378 -619.076 374.299 -614.047 370.703 -608.9 367.225 -603.625 363.875 -598.209 360.666 -592.641 357.609 -589.796 356.142 -586.908 354.717 -583.977 353.336 -581 352 -578.937 351.098 -576.871 350.27 -575.836 349.884 -574.799 349.517 -573.76 349.17 -572.719 348.844 -571.674 348.538 -570.626 348.252 -569.575 347.988 -568.52 347.746 -567.46 347.526 -566.395 347.328 -565.325 347.152 -564.25 347 -563.169 346.871 -562.082 346.766 -560.988 346.685 -559.887 346.629 -558.779 346.597 -557.663 346.591 -556.539 346.611 -555.406 346.656 -554.265 346.728 -553.115 346.827 -551.955 346.952 -550.785 347.105 -549.605 347.286 -548.415 347.496 -547.213 347.733 -546 348 -545.626 348.074 -545.254 348.155 -544.885 348.244 -544.518 348.339 -544.154 348.441 -543.791 348.551 -543.432 348.667 -543.075 348.79 -542.72 348.919 -542.368 349.056 -542.019 349.198 -541.673 349.347 -541.33 349.503 -540.989 349.664 -540.651 349.832 -540.316 350.006 -539.985 350.186 -539.656 350.372 -539.33 350.563 -539.008 350.76 -538.373 351.172 -537.751 351.606 -537.144 352.061 -536.55 352.536 -535.971 353.032 -535.406 353.547 -534.857 354.081 -534.324 354.632 -533.807 355.201 -533.307 355.787 -532.824 356.389 -532.358 357.006 -531.91 357.637 -531.48 358.283 -531.069 358.942 -530.678 359.614 -530.305 360.298 -529.953 360.993 -529.62 361.699 -529.309 362.415 -529.019 363.141 -528.75 363.875 -528.503 364.617 -528.279 365.367 -528.077 366.123 -527.899 366.885 -527.744 367.653 -527.613 368.426 -527.507 369.202 -527.426 369.982 -527.37 370.765 -527.339 371.55 -527.335 372.336 -527.357 373.123 -527.406 373.91 -527.441 374.304 -527.482 374.697 -527.531 375.09 -527.586 375.482 -527.649 375.874 -527.719 376.266 -527.796 376.656 -527.88 377.046 -527.971 377.436 -528.069 377.824 -528.175 378.211 -528.289 378.598 -528.409 378.983 -528.538 379.367 -528.673 379.75 -528.817 380.131 -528.968 380.511 -529.127 380.889 -529.293 381.266 -529.468 381.64 -529.65 382.014 -529.84 382.385 -530.038 382.754 -530.244 383.121 -530.458 383.486 -530.68 383.849 -530.911 384.21 -531.15 384.568 -531.397 384.923 -531.652 385.277 -531.916 385.627 -532.188 385.975 -532.468 386.32 -532.757 386.662 -533.055 387.001 -533.361 387.337 -533.676 387.67 -534 388 -535.222 389.107 -536.45 390.18 -537.684 391.218 -538.926 392.223 -540.175 393.193 -541.431 394.13 -542.696 395.035 -543.969 395.906 -545.251 396.746 -546.542 397.553 -547.842 398.329 -549.152 399.074 -550.473 399.788 -551.804 400.472 -553.146 401.126 -554.5 401.75 -555.865 402.345 -557.243 402.911 -558.633 403.448 -560.035 403.957 -561.451 404.438 -562.88 404.892 -564.324 405.319 -565.781 405.719 -567.253 406.092 -568.741 406.44 -570.243 406.762 -571.762 407.059 -573.296 407.33 -574.847 407.578 -578 408 -581.949 408.328 -585.918 408.564 -589.905 408.71 -593.906 408.766 -597.919 408.734 -601.941 408.615 -605.969 408.412 -610 408.125 -614.031 407.756 -618.059 407.307 -622.081 406.778 -626.094 406.172 -630.095 405.49 -634.082 404.732 -638.051 403.902 -642 403 -642.284 402.909 -642.574 402.824 -642.869 402.743 -643.17 402.668 -643.786 402.53 -644.422 402.406 -645.075 402.294 -645.744 402.191 -647.125 402 -648.553 401.809 -650.016 401.594 -650.757 401.47 -651.502 401.332 -652.25 401.176 -653 401 m -405 474 l -404.549 472.972 -404.132 471.949 -403.749 470.931 -403.4 469.918 -403.084 468.909 -402.801 467.905 -402.549 466.904 -402.328 465.906 -402.138 464.912 -401.979 463.919 -401.849 462.93 -401.748 461.941 -401.676 460.955 -401.632 459.969 -401.615 458.984 -401.625 458 -401.662 457.016 -401.724 456.031 -401.811 455.045 -401.924 454.059 -402.06 453.07 -402.22 452.081 -402.404 451.088 -402.609 450.094 -402.837 449.096 -403.086 448.095 -403.356 447.091 -403.646 446.082 -404.286 444.051 -405 442 -405.27 441.062 -405.516 440.125 -405.74 439.188 -405.941 438.25 -406.121 437.312 -406.279 436.375 -406.415 435.438 -406.531 434.5 -406.627 433.562 -406.702 432.625 -406.757 431.688 -406.793 430.75 -406.81 429.812 -406.808 428.875 -406.788 427.938 -406.75 427 -406.694 426.062 -406.622 425.125 -406.532 424.188 -406.426 423.25 -406.166 421.375 -405.844 419.5 -405.463 417.625 -405.027 415.75 -404.539 413.875 -404 412 -403.903 411.725 -403.801 411.46 -403.747 411.333 -403.692 411.207 -403.636 411.085 -403.578 410.965 -403.519 410.847 -403.458 410.733 -403.396 410.62 -403.332 410.51 -403.267 410.403 -403.2 410.297 -403.132 410.194 -403.062 410.094 -402.991 409.995 -402.919 409.899 -402.845 409.805 -402.77 409.712 -402.693 409.622 -402.614 409.534 -402.534 409.448 -402.453 409.363 -402.37 409.281 -402.286 409.2 -402.2 409.121 -402.113 409.043 -402.025 408.968 -401.935 408.894 -401.843 408.821 -401.75 408.75 -401.656 408.68 -401.56 408.612 -401.363 408.48 -401.161 408.353 -400.953 408.23 -400.739 408.112 -400.52 407.999 -400.294 407.888 -400.062 407.781 -399.825 407.677 -399.582 407.576 -399.078 407.379 -398.551 407.188 -398 407 -395.398 405.84 -392.844 404.611 -390.336 403.315 -387.875 401.953 -385.461 400.527 -383.094 399.037 -380.773 397.486 -378.5 395.875 -376.273 394.205 -374.094 392.479 -371.961 390.696 -369.875 388.859 -367.836 386.97 -365.844 385.029 -363.898 383.039 -362 381 -361.906 380.862 -361.812 380.73 -361.719 380.604 -361.625 380.483 -361.531 380.368 -361.438 380.258 -361.344 380.154 -361.25 380.055 -361.156 379.961 -361.062 379.871 -360.969 379.787 -360.875 379.708 -360.781 379.633 -360.688 379.564 -360.594 379.498 -360.5 379.438 -360.406 379.381 -360.312 379.329 -360.266 379.305 -360.219 379.281 -360.172 379.259 -360.125 379.237 -360.078 379.217 -360.031 379.198 -359.984 379.179 -359.938 379.162 -359.891 379.145 -359.844 379.13 -359.797 379.115 -359.75 379.102 -359.703 379.089 -359.656 379.077 -359.609 379.066 -359.562 379.056 -359.516 379.047 -359.469 379.039 -359.422 379.031 -359.375 379.024 -359.281 379.014 -359.188 379.006 -359.094 379.001 -359 379 -358.906 379.001 -358.812 379.006 -358.719 379.013 -358.625 379.022 -358.531 379.035 -358.438 379.049 -358.344 379.067 -358.25 379.086 -358.156 379.108 -358.062 379.131 -357.969 379.157 -357.875 379.185 -357.781 379.214 -357.688 379.245 -357.5 379.312 -357.312 379.386 -357.125 379.464 -356.938 379.547 -356.75 379.633 -356.375 379.813 -356 380 -355.536 380.379 -355.083 380.767 -354.64 381.164 -354.207 381.569 -353.784 381.982 -353.372 382.403 -352.97 382.832 -352.578 383.27 -352.197 383.714 -351.825 384.167 -351.464 384.626 -351.113 385.093 -350.773 385.567 -350.442 386.049 -350.122 386.537 -349.812 387.031 -349.513 387.533 -349.224 388.04 -348.945 388.554 -348.676 389.075 -348.417 389.601 -348.169 390.133 -347.931 390.671 -347.703 391.215 -347.486 391.764 -347.278 392.318 -347.081 392.878 -346.895 393.443 -346.718 394.013 -346.552 394.587 -346.396 395.166 -346.25 395.75 -346.115 396.338 -345.989 396.93 -345.874 397.527 -345.77 398.127 -345.675 398.732 -345.591 399.34 -345.453 400.566 -345.356 401.806 -345.301 403.058 -345.286 404.321 -345.312 405.594 -345.38 406.875 -345.488 408.165 -345.638 409.46 -345.828 410.762 -346.06 412.067 -346.332 413.376 -346.646 414.688 -347 416 -347.29 416.932 -347.598 417.851 -347.923 418.759 -348.266 419.654 -348.626 420.537 -349.004 421.407 -349.399 422.265 -349.812 423.109 -350.243 423.941 -350.691 424.759 -351.157 425.563 -351.641 426.354 -352.142 427.13 -352.66 427.893 -353.196 428.641 -353.75 429.375 -354.321 430.094 -354.91 430.799 -355.517 431.488 -356.141 432.162 -356.782 432.821 -357.441 433.464 -358.118 434.092 -358.812 434.703 -359.524 435.299 -360.254 435.878 -361.001 436.44 -361.766 436.986 -362.548 437.515 -363.348 438.028 -364.165 438.522 -365 439 -366.301 439.762 -367.579 440.546 -368.836 441.351 -370.072 442.178 -371.289 443.024 -372.486 443.889 -374.828 445.672 -377.106 447.519 -379.326 449.424 -381.497 451.378 -383.625 453.375 -387.783 457.467 -391.859 461.641 -395.912 465.838 -400 470 -400.048 470.093 -400.097 470.185 -400.147 470.275 -400.199 470.364 -400.306 470.538 -400.418 470.707 -400.534 470.872 -400.655 471.033 -400.779 471.19 -400.906 471.344 -401.037 471.495 -401.169 471.643 -401.441 471.934 -401.719 472.218 -402 472.5 -402.281 472.782 -402.559 473.066 -402.831 473.357 -402.963 473.505 -403.094 473.656 -403.221 473.81 -403.345 473.967 -403.466 474.128 -403.582 474.293 -403.694 474.462 -403.801 474.636 -403.853 474.725 -403.903 474.815 -403.952 474.907 -404 475 -405 474 m -315 68 l -315.009 68.747 -315.034 69.488 -315.077 70.224 -315.135 70.953 -315.208 71.677 -315.297 72.395 -315.399 73.106 -315.516 73.812 -315.645 74.513 -315.787 75.207 -315.942 75.896 -316.107 76.578 -316.471 77.926 -316.875 79.25 -317.314 80.551 -317.783 81.828 -318.279 83.082 -318.797 84.312 -319.332 85.52 -319.881 86.703 -321 89 -322.502 91.91 -324.016 94.656 -325.553 97.262 -327.125 99.75 -328.744 102.145 -330.422 104.469 -332.17 106.746 -334 109 -334.707 109.701 -335.033 110.013 -335.344 110.297 -335.495 110.428 -335.643 110.552 -335.789 110.667 -335.934 110.775 -336.005 110.826 -336.076 110.875 -336.147 110.922 -336.218 110.967 -336.289 111.01 -336.359 111.05 -336.43 111.089 -336.5 111.125 -336.57 111.159 -336.641 111.191 -336.711 111.22 -336.782 111.248 -336.853 111.273 -336.924 111.296 -336.995 111.316 -337.066 111.334 -337.138 111.35 -337.211 111.363 -337.284 111.374 -337.357 111.382 -337.431 111.388 -337.505 111.391 -337.58 111.392 -337.656 111.391 -337.733 111.386 -337.81 111.38 -337.888 111.37 -337.967 111.358 -338.047 111.343 -338.128 111.326 -338.21 111.306 -338.293 111.283 -338.377 111.258 -338.462 111.229 -338.549 111.198 -338.636 111.164 -338.725 111.128 -338.815 111.088 -338.907 111.045 -339 111 -340.102 110.239 -341.158 109.457 -342.171 108.658 -343.141 107.844 -344.07 107.018 -344.959 106.184 -345.81 105.343 -346.625 104.5 -347.405 103.657 -348.15 102.816 -348.864 101.982 -349.547 101.156 -350.826 99.543 -352 98 -352.091 97.907 -352.176 97.815 -352.255 97.725 -352.329 97.636 -352.398 97.549 -352.461 97.463 -352.519 97.378 -352.546 97.336 -352.572 97.295 -352.597 97.254 -352.62 97.213 -352.643 97.172 -352.664 97.132 -352.683 97.092 -352.702 97.052 -352.72 97.013 -352.736 96.974 -352.751 96.935 -352.766 96.897 -352.779 96.858 -352.791 96.821 -352.802 96.783 -352.811 96.746 -352.82 96.709 -352.828 96.672 -352.835 96.635 -352.841 96.599 -352.846 96.563 -352.85 96.528 -352.852 96.492 -352.854 96.457 -352.856 96.422 -352.856 96.387 -352.855 96.353 -352.853 96.319 -352.851 96.285 -352.848 96.251 -352.843 96.218 -352.839 96.185 -352.833 96.152 -352.826 96.119 -352.819 96.087 -352.811 96.054 -352.802 96.022 -352.792 95.991 -352.782 95.959 -352.771 95.928 -352.759 95.896 -352.747 95.865 -352.734 95.835 -352.72 95.804 -352.706 95.774 -352.691 95.744 -352.675 95.714 -352.659 95.684 -352.642 95.654 -352.625 95.625 -352.589 95.567 -352.55 95.509 -352.51 95.452 -352.468 95.396 -352.424 95.341 -352.379 95.286 -352.332 95.231 -352.283 95.178 -352.234 95.125 -352.183 95.072 -352.078 94.968 -351.97 94.866 -351.859 94.766 -351.635 94.569 -351.412 94.377 -351.198 94.188 -351.097 94.094 -351 94 -350.259 93.338 -349.535 92.665 -348.827 91.98 -348.137 91.285 -347.462 90.58 -346.803 89.865 -346.16 89.14 -345.531 88.406 -344.917 87.664 -344.318 86.913 -343.16 85.387 -342.055 83.832 -341 82.25 -339.992 80.645 -339.027 79.02 -338.104 77.376 -337.219 75.719 -336.369 74.049 -335.551 72.371 -334 69 -333.254 66.707 -332.531 64.344 -331.855 61.934 -331.542 60.718 -331.25 59.5 -330.981 58.282 -330.738 57.066 -330.525 55.857 -330.344 54.656 -330.198 53.467 -330.09 52.293 -330.051 51.712 -330.023 51.136 -330.006 50.565 -330 50 -329.997 49.727 -329.988 49.472 -329.974 49.234 -329.953 49.012 -329.941 48.907 -329.927 48.806 -329.911 48.709 -329.895 48.616 -329.876 48.526 -329.856 48.441 -329.835 48.359 -329.812 48.281 -329.788 48.207 -329.776 48.171 -329.763 48.136 -329.749 48.102 -329.736 48.069 -329.721 48.037 -329.707 48.005 -329.692 47.975 -329.677 47.945 -329.661 47.916 -329.646 47.888 -329.629 47.861 -329.613 47.835 -329.596 47.81 -329.578 47.785 -329.56 47.761 -329.542 47.738 -329.524 47.716 -329.505 47.695 -329.486 47.674 -329.466 47.655 -329.446 47.636 -329.426 47.618 -329.405 47.6 -329.384 47.584 -329.363 47.568 -329.341 47.553 -329.319 47.538 -329.296 47.525 -329.273 47.512 -329.25 47.5 -329.226 47.489 -329.202 47.478 -329.178 47.468 -329.153 47.459 -329.128 47.45 -329.103 47.443 -329.077 47.436 -329.051 47.429 -329.024 47.423 -328.997 47.418 -328.97 47.414 -328.942 47.41 -328.914 47.407 -328.886 47.405 -328.857 47.403 -328.828 47.402 -328.799 47.402 -328.769 47.402 -328.739 47.403 -328.708 47.404 -328.677 47.407 -328.646 47.409 -328.614 47.413 -328.582 47.417 -328.517 47.426 -328.45 47.438 -328.382 47.452 -328.312 47.469 -328.241 47.487 -328.169 47.508 -328.095 47.531 -328.02 47.556 -327.864 47.612 -327.703 47.676 -327.536 47.747 -327.363 47.825 -327.185 47.909 -327 48 -326.262 48.384 -325.547 48.785 -324.855 49.203 -324.188 49.639 -323.543 50.091 -323.229 50.323 -322.922 50.56 -322.62 50.8 -322.324 51.045 -322.034 51.294 -321.75 51.547 -321.472 51.804 -321.199 52.065 -320.933 52.33 -320.672 52.598 -320.417 52.871 -320.168 53.148 -319.925 53.428 -319.688 53.713 -319.456 54.001 -319.23 54.293 -319.011 54.589 -318.797 54.889 -318.589 55.192 -318.387 55.5 -318.19 55.81 -318 56.125 -317.815 56.443 -317.637 56.765 -317.464 57.091 -317.297 57.42 -317.136 57.752 -316.98 58.089 -316.831 58.428 -316.688 58.771 -316.55 59.118 -316.418 59.468 -316.292 59.822 -316.172 60.179 -316.058 60.539 -315.949 60.903 -315.75 61.641 -315.574 62.391 -315.422 63.155 -315.293 63.932 -315.188 64.721 -315.105 65.522 -315.047 66.336 -315.012 67.162 -315 68 m -360 83 l -360.902 81.312 -361.734 79.623 -362.124 78.777 -362.496 77.931 -362.851 77.083 -363.188 76.234 -363.507 75.384 -363.809 74.532 -364.093 73.678 -364.359 72.822 -364.608 71.964 -364.84 71.104 -365.054 70.241 -365.25 69.375 -365.429 68.506 -365.59 67.635 -365.733 66.759 -365.859 65.881 -365.968 64.999 -366.059 64.113 -366.132 63.222 -366.188 62.328 -366.226 61.429 -366.246 60.526 -366.249 59.618 -366.234 58.705 -366.202 57.787 -366.152 56.864 -366.085 55.935 -366 55 -365.984 49.906 -365.947 47.441 -365.875 45 -365.756 42.559 -365.578 40.094 -365.33 37.582 -365 35 -364.812 34.461 -364.718 34.209 -364.623 33.969 -364.527 33.74 -364.431 33.523 -364.333 33.318 -364.234 33.125 -364.184 33.033 -364.134 32.943 -364.083 32.857 -364.032 32.773 -363.98 32.693 -363.928 32.615 -363.875 32.541 -363.822 32.469 -363.769 32.4 -363.714 32.334 -363.659 32.271 -363.604 32.211 -363.548 32.154 -363.491 32.1 -363.433 32.048 -363.404 32.024 -363.375 32 -363.346 31.977 -363.316 31.955 -363.286 31.933 -363.256 31.912 -363.226 31.892 -363.196 31.873 -363.165 31.854 -363.135 31.836 -363.104 31.819 -363.072 31.802 -363.041 31.786 -363.009 31.771 -362.978 31.757 -362.946 31.744 -362.913 31.731 -362.881 31.719 -362.848 31.707 -362.815 31.697 -362.782 31.687 -362.749 31.678 -362.715 31.669 -362.681 31.662 -362.647 31.655 -362.613 31.648 -362.578 31.643 -362.543 31.638 -362.508 31.634 -362.472 31.631 -362.437 31.628 -362.401 31.626 -362.365 31.625 -362.328 31.625 -362.291 31.625 -362.254 31.626 -362.217 31.628 -362.179 31.631 -362.142 31.634 -362.103 31.638 -362.065 31.643 -362.026 31.648 -361.948 31.662 -361.868 31.678 -361.787 31.697 -361.705 31.719 -361.622 31.744 -361.537 31.771 -361.451 31.802 -361.364 31.836 -361.275 31.873 -361.185 31.912 -361.093 31.955 -361 32 -360.906 32.046 -360.81 32.091 -360.614 32.176 -360.413 32.257 -360.207 32.332 -359.997 32.403 -359.783 32.47 -359.565 32.534 -359.344 32.594 -358.893 32.706 -358.434 32.809 -357.5 33 -356.566 33.191 -356.107 33.294 -355.656 33.406 -355.435 33.466 -355.217 33.53 -355.003 33.597 -354.793 33.668 -354.587 33.743 -354.386 33.824 -354.19 33.909 -354.094 33.954 -354 34 -353.262 34.381 -352.547 34.774 -351.856 35.179 -351.189 35.596 -350.547 36.025 -349.928 36.468 -349.335 36.923 -349.047 37.155 -348.766 37.391 -348.49 37.63 -348.221 37.872 -347.959 38.117 -347.702 38.366 -347.452 38.619 -347.209 38.875 -346.971 39.134 -346.74 39.396 -346.516 39.663 -346.298 39.932 -346.086 40.205 -345.881 40.482 -345.682 40.762 -345.49 41.046 -345.304 41.334 -345.125 41.625 -344.953 41.92 -344.787 42.218 -344.627 42.52 -344.475 42.826 -344.329 43.136 -344.19 43.45 -344.057 43.767 -343.932 44.088 -343.813 44.413 -343.701 44.742 -343.595 45.074 -343.497 45.411 -343.405 45.751 -343.321 46.096 -343.243 46.444 -343.172 46.797 -343.108 47.153 -343.051 47.514 -343.001 47.879 -342.958 48.247 -342.922 48.62 -342.894 48.997 -342.872 49.378 -342.857 49.764 -342.85 50.153 -342.85 50.547 -342.857 50.945 -342.871 51.347 -342.921 52.165 -343 53 -343.202 54.116 -343.433 55.216 -343.691 56.298 -343.977 57.365 -344.288 58.417 -344.626 59.453 -344.988 60.476 -345.375 61.484 -345.785 62.48 -346.218 63.463 -346.672 64.433 -347.148 65.393 -347.645 66.341 -348.161 67.279 -348.696 68.206 -349.25 69.125 -349.821 70.035 -350.409 70.936 -351.633 72.717 -352.915 74.471 -354.25 76.203 -355.632 77.918 -357.055 79.619 -360 83 m -562 292 l -561.824 292.75 -561.672 293.5 -561.543 294.25 -561.438 295 -561.355 295.75 -561.297 296.5 -561.262 297.25 -561.25 298 -561.262 298.75 -561.297 299.5 -561.355 300.25 -561.438 301 -561.543 301.75 -561.672 302.5 -561.824 303.25 -562 304 -562.176 303.25 -562.328 302.5 -562.457 301.75 -562.562 301 -562.645 300.25 -562.703 299.5 -562.738 298.75 -562.75 298 -562.738 297.25 -562.703 296.5 -562.645 295.75 -562.562 295 -562.457 294.25 -562.328 293.5 -562.176 292.75 -562 292 m -158 304 l -158.182 303.628 -158.352 303.261 -158.51 302.901 -158.656 302.545 -158.791 302.194 -158.914 301.849 -159.025 301.508 -159.125 301.172 -159.213 300.84 -159.289 300.512 -159.354 300.189 -159.406 299.869 -159.428 299.711 -159.447 299.553 -159.463 299.396 -159.477 299.24 -159.487 299.085 -159.494 298.931 -159.499 298.778 -159.5 298.625 -159.499 298.473 -159.494 298.322 -159.487 298.171 -159.477 298.021 -159.463 297.872 -159.447 297.723 -159.428 297.575 -159.406 297.428 -159.381 297.281 -159.354 297.134 -159.323 296.988 -159.289 296.843 -159.252 296.698 -159.213 296.553 -159.17 296.409 -159.125 296.266 -159.077 296.122 -159.025 295.979 -158.971 295.837 -158.914 295.694 -158.854 295.552 -158.791 295.41 -158.725 295.268 -158.656 295.127 -158.51 294.845 -158.352 294.563 -158.182 294.281 -158 294 -157.909 294.281 -157.824 294.563 -157.745 294.845 -157.672 295.127 -157.604 295.41 -157.543 295.694 -157.487 295.979 -157.438 296.266 -157.394 296.553 -157.355 296.843 -157.323 297.134 -157.297 297.428 -157.276 297.723 -157.262 298.021 -157.253 298.322 -157.25 298.625 -157.253 298.931 -157.262 299.24 -157.276 299.553 -157.297 299.869 -157.355 300.512 -157.438 301.172 -157.543 301.849 -157.672 302.545 -157.824 303.261 -158 304 m -415 82 l -415.182 81.719 -415.352 81.438 -415.511 81.156 -415.658 80.875 -415.795 80.594 -415.921 80.312 -416.036 80.031 -416.141 79.75 -416.235 79.469 -416.32 79.188 -416.394 78.906 -416.459 78.625 -416.514 78.344 -416.56 78.062 -416.597 77.781 -416.625 77.5 -416.644 77.219 -416.655 76.938 -416.657 76.656 -416.65 76.375 -416.636 76.094 -416.614 75.812 -416.584 75.531 -416.547 75.25 -416.502 74.969 -416.45 74.688 -416.392 74.406 -416.326 74.125 -416.254 73.844 -416.176 73.562 -416.091 73.281 -416 73 -415.909 73.281 -415.824 73.562 -415.744 73.844 -415.67 74.125 -415.536 74.688 -415.422 75.25 -415.325 75.812 -415.244 76.375 -415.178 76.938 -415.125 77.5 -415.084 78.062 -415.053 78.625 -415.016 79.75 -415.002 80.875 -415 82 m -378 37 l -378.002 36.627 -378.007 36.444 -378.016 36.266 -378.022 36.178 -378.031 36.093 -378.041 36.009 -378.053 35.928 -378.067 35.848 -378.075 35.809 -378.084 35.771 -378.093 35.734 -378.103 35.697 -378.114 35.661 -378.125 35.625 -378.137 35.59 -378.15 35.556 -378.164 35.523 -378.178 35.49 -378.193 35.459 -378.209 35.428 -378.218 35.413 -378.226 35.398 -378.235 35.384 -378.244 35.369 -378.253 35.355 -378.263 35.341 -378.273 35.327 -378.283 35.314 -378.293 35.301 -378.303 35.288 -378.314 35.275 -378.325 35.262 -378.336 35.25 -378.348 35.238 -378.359 35.227 -378.371 35.215 -378.384 35.204 -378.396 35.193 -378.409 35.182 -378.422 35.172 -378.435 35.162 -378.449 35.152 -378.463 35.142 -378.477 35.133 -378.491 35.124 -378.506 35.115 -378.521 35.107 -378.536 35.099 -378.552 35.091 -378.568 35.084 -378.584 35.076 -378.601 35.069 -378.618 35.063 -378.635 35.057 -378.652 35.051 -378.67 35.045 -378.688 35.04 -378.706 35.035 -378.725 35.03 -378.744 35.026 -378.764 35.022 -378.783 35.018 -378.804 35.014 -378.824 35.011 -378.845 35.009 -378.866 35.006 -378.887 35.005 -378.909 35.003 -378.954 35.001 -379 35 -378.954 34.999 -378.909 34.997 -378.866 34.994 -378.824 34.989 -378.783 34.983 -378.743 34.976 -378.705 34.968 -378.668 34.959 -378.632 34.949 -378.597 34.938 -378.563 34.927 -378.53 34.914 -378.498 34.901 -378.466 34.888 -378.436 34.874 -378.406 34.859 -378.377 34.845 -378.349 34.829 -378.294 34.799 -378.242 34.767 -378.191 34.736 -378.142 34.706 -378.094 34.677 -378.071 34.663 -378.047 34.65 -378.023 34.637 -378 34.625 -377.977 34.614 -377.953 34.603 -377.941 34.598 -377.929 34.593 -377.918 34.589 -377.906 34.585 -377.894 34.581 -377.882 34.577 -377.87 34.574 -377.858 34.57 -377.846 34.567 -377.833 34.565 -377.821 34.563 -377.809 34.561 -377.796 34.559 -377.784 34.558 -377.771 34.557 -377.758 34.556 -377.745 34.556 -377.732 34.556 -377.719 34.556 -377.706 34.557 -377.692 34.558 -377.679 34.56 -377.665 34.562 -377.651 34.564 -377.637 34.567 -377.623 34.57 -377.608 34.574 -377.594 34.578 -377.579 34.583 -377.564 34.588 -377.549 34.593 -377.534 34.599 -377.518 34.606 -377.502 34.613 -377.486 34.621 -377.47 34.629 -377.454 34.637 -377.437 34.646 -377.42 34.656 -377.403 34.666 -377.368 34.688 -377.332 34.713 -377.295 34.74 -377.257 34.769 -377.217 34.801 -377.176 34.835 -377.134 34.872 -377.091 34.912 -377 35 -377.002 35.373 -377.007 35.556 -377.016 35.734 -377.022 35.822 -377.031 35.907 -377.041 35.991 -377.053 36.072 -377.067 36.152 -377.075 36.191 -377.084 36.229 -377.093 36.266 -377.103 36.303 -377.114 36.339 -377.125 36.375 -377.137 36.41 -377.15 36.444 -377.164 36.477 -377.178 36.51 -377.193 36.541 -377.209 36.572 -377.218 36.587 -377.226 36.602 -377.235 36.616 -377.244 36.631 -377.253 36.645 -377.263 36.659 -377.273 36.673 -377.283 36.686 -377.293 36.699 -377.303 36.712 -377.314 36.725 -377.325 36.738 -377.336 36.75 -377.348 36.762 -377.359 36.773 -377.371 36.785 -377.384 36.796 -377.396 36.807 -377.409 36.818 -377.422 36.828 -377.435 36.838 -377.449 36.848 -377.463 36.858 -377.477 36.867 -377.491 36.876 -377.506 36.885 -377.521 36.893 -377.536 36.901 -377.552 36.909 -377.568 36.916 -377.584 36.924 -377.601 36.931 -377.618 36.937 -377.635 36.943 -377.652 36.949 -377.67 36.955 -377.688 36.96 -377.706 36.965 -377.725 36.97 -377.744 36.974 -377.764 36.978 -377.783 36.982 -377.804 36.986 -377.824 36.989 -377.845 36.991 -377.866 36.994 -377.887 36.995 -377.909 36.997 -377.954 36.999 -378 37"
+
+TEST_SPLIT_COMPLEX1_ORIGINAL = "m -100.5 0 l 100 0 b 100 100 -100 100 -100.5 0 c"
+TEST_SPLIT_COMPLEX1_DEST = "m -100.5 0 l -92 0 -76 0 -60 0 -44 0 -28 0 -12 0 4 0 20 0 36 0 52 0 68 0 84 0 100 0 99.964 2.325 99.855 4.614 99.676 6.866 99.426 9.082 99.108 11.261 98.723 13.403 98.271 15.509 97.754 17.578 97.173 19.611 96.528 21.606 95.822 23.566 95.056 25.488 94.23 27.374 93.345 29.224 92.403 31.036 91.405 32.812 90.352 34.552 89.246 36.255 88.086 37.921 86.876 39.551 85.614 41.144 84.304 42.7 82.945 44.22 81.54 45.703 80.088 47.15 78.592 48.56 77.053 49.933 75.471 51.27 73.848 52.57 72.184 53.833 70.482 55.06 68.742 56.25 66.965 57.404 65.153 58.521 63.307 59.601 61.427 60.645 59.515 61.652 57.572 62.622 55.599 63.556 53.598 64.453 51.569 65.314 49.514 66.138 47.433 66.925 45.329 67.676 43.201 68.39 41.052 69.067 38.882 69.708 36.692 70.312 34.484 70.88 32.259 71.411 27.762 72.363 23.209 73.169 18.61 73.828 13.975 74.341 9.311 74.707 4.629 74.927 -0.062 75 -4.755 74.927 -9.438 74.707 -14.103 74.341 -18.741 73.828 -23.343 73.169 -27.9 72.363 -32.402 71.411 -34.63 70.88 -36.841 70.312 -39.033 69.708 -41.207 69.067 -43.359 68.39 -45.49 67.676 -47.599 66.925 -49.683 66.138 -51.743 65.314 -53.776 64.453 -55.782 63.556 -57.759 62.622 -59.707 61.652 -61.624 60.645 -63.509 59.601 -65.361 58.521 -67.178 57.404 -68.961 56.25 -70.707 55.06 -72.415 53.833 -74.085 52.57 -75.714 51.27 -77.303 49.933 -78.85 48.56 -80.353 47.15 -81.811 45.703 -83.224 44.22 -84.59 42.7 -85.909 41.144 -87.178 39.551 -88.397 37.921 -89.564 36.255 -90.68 34.552 -91.741 32.812 -92.748 31.036 -93.699 29.224 -94.593 27.374 -95.428 25.488 -96.205 23.566 -96.92 21.606 -97.575 19.611 -98.166 17.578 -98.693 15.509 -99.156 13.403 -99.552 11.261 -99.881 9.082 -100.141 6.866 -100.332 4.614 -100.452 2.325 -100.5 0"
